% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{ask_password}
\alias{ask_password}
\title{Get password from the user if \code{\link[=interactive]{interactive()}}}
\usage{
ask_password()
}
\value{
Invisibly returns a string if \code{\link[=interactive]{interactive()}}. If it is not an interactive R
session, then a \code{\link[=stop]{stop()}} message is displayed on the console.
}
\description{
Get password from the user if \code{\link[=interactive]{interactive()}}
}
\author{
Kelli F. Johnson
}
