% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capValues.R
\name{capValues}
\alias{capValues}
\title{Decrease all values greater than a maximum specified value.}
\usage{
capValues(DataCol, maxVal = 0.95)
}
\arguments{
\item{DataCol}{A vector of numeric values.}

\item{maxVal}{A numeric value specifying the maximum value or quantile
at which to cap all data.}
}
\value{
A vector of numeric values where no value is greater than the
maximum specified by \code{maxVal}.
}
\description{
\code{capValues} takes a numeric vector and returns a vector in which all
values greater than a specified maximum are reset to that maximum
value.
}
\details{
The maximum may be specified either as a
quantile or as a number.

If \code{maxVal} is less than one, it is
interpreted as a quantile, otherwise it is interpreted as the maximum
value to return.

The default value is \code{0.95}, which caps all
values in the input at the 95 percent quantile.
}
\section{Example}{


x = seq(1,10)
y <- capValues(x, 0.75)
z <- capValues(x, 5)
rbind(x, y, z)
}

\author{
Andi Stephens
}
