\name{Catch.XMPL}
\alias{Catch.XMPL}
\docType{data}
\title{
Sample dataset for Catch
}
\description{
Example of Catch data for level-2 expansions.
}
\usage{data("Catch.XMPL")}
\format{
  A data frame with 16 observations on the following 7 variables.
  \describe{
    \item{\code{Year}}{a numeric vector}
    \item{\code{CA.HKL}}{a numeric vector}
    \item{\code{CA.TWL}}{a numeric vector}
    \item{\code{OR.HKL}}{a numeric vector}
    \item{\code{OR.TWL}}{a numeric vector}
    \item{\code{WA.HKL}}{a numeric vector}
    \item{\code{WA.TWL}}{a numeric vector}
  }
}
\examples{
data(Catch.XMPL)
## maybe str(Catch.XMPL) ; plot(Catch.XMPL) ...
}
\keyword{datasets}
