% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_representativeness_plot.R
\name{changecol_pacfin}
\alias{changecol_pacfin}
\title{Change Column Names of PacFIN Data To Match Survey}
\usage{
changecol_pacfin(data, verbose = TRUE)
}
\arguments{
\item{data}{A data frame with names matching those from PacFIN.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
A data frame with names similar to those exported by the survey-data
warehouse. No column are actually removed, columns are only added.
}
\description{
Change the column names of data extracted from PacFIN to match those
extracted from the survey-data warehouse.
}
