% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkINPFC.R
\name{checkINPFC}
\alias{checkINPFC}
\title{Check  International North Pacific Fisheries Commission (INPFC) Areas}
\usage{
checkINPFC(data, keep = ls_INPFC(area = "US"))
}
\arguments{
\item{data}{A data frame with named columns.}

\item{keep}{A vector of types to keep. See \code{formals(checkINPFC)$keep}
for default values.}
}
\value{
A vector of TRUE/FALSE values indicating which INPFC areas
are in those specified in the \code{keep} argument.
}
\description{
Check that the INPFC areas are valid and within the areas to keep.
Note that many PacFIN records do not have an entry for \code{INPFC_AREA},
and thus, if you wish to keep these, then \code{""} will need to be added
to the vector supplied to the \code{"keep"} argument.
}
\examples{
checkINPFC(1:10, keep = ls_INPFC(area = "All"))
\dontshow{
testthat::expect_equal(
  c(TRUE, TRUE, TRUE, FALSE),
  checkINPFC(c("CAN", "CT", "GS", "AAA"), keep = ls_INPFC(area = "All"))
)
}
}
