% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLenAge.R
\name{checkLenAge}
\alias{checkLenAge}
\title{Calculate the von Bertanlaffy growth funciton and compare length
age value to those estimated with the option to throw out records
that are outside a specified sd band.}
\usage{
checkLenAge(
  Pdata,
  Par = list(0.13, 55, 15, 0.1, 0.1),
  keepAll = TRUE,
  sdFactor = 4,
  Optim = TRUE
)
}
\arguments{
\item{Pdata}{a PacFIN dataset}

\item{Par}{initial or externally estimated parameter values for the vonB
growth function where the order are ("k", "Linf", "Lmin", "CV young", "CV old").
Can define a list with multiple values by sex where the expected order is 
c(k.vec, Linf.vec, Lmin.vec, CVyoung.vec, CVold.vec) where each vector is c(female, male, unsex)
parameter value.
The lengths are input in cms for ease of use and are converted internally to mm.}

\item{keepAll}{Option to NA inconsistent lengths}

\item{sdFactor}{Option to control the threshold of the estimated low and high bound
that is compared to when throwing out data. Default value is 4 stan. devs. away from the 
predictions.}

\item{Optim}{Control whether to internally solve for the parameter estimates or use the input
parameters.}
}
\value{
Returns a combined dataset in PacFIN format.
}
\description{
\subsection{Workflow}{
Intended to be run after \code{\link{cleanPacFIN}}.
}
}
\details{

}
\author{
Chantel Wetzel, Vlada Gertseva, James Thorson
}
