% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkparam.R
\name{checkparam.Comps}
\alias{checkparam.Comps}
\title{Check Argument Comps}
\usage{
checkparam.Comps(Comps = c("LEN", "AGE", "AAL", "NA"))
}
\arguments{
\item{Comps}{The type of composition data to create. Options are
length (\code{"LEN"},
age (\code{"AGE"}),
or conditional age-at-length (\code{"AAL"}).
For some functions, a value of \code{NA} will allow for non-specific
code to be run, i.e., not based on the composition type.
Note that the entry needs to be in all upper-case letters.}
}
\value{
A single character value.
}
\description{
Check and select the appropriate input value for the Comps
parameter used in multiple functions within \pkg{PacFIN.Utilities}.
}
\author{
Kelli F. Johnson
}
