% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanAges.R
\name{cleanAges}
\alias{cleanAges}
\title{Filter samples for appropriate ages and agemethods.}
\usage{
cleanAges(Pdata, keep_age_methods = c("B", "S", ""), minAge = 1,
  maxAge = NULL, CLEAN = TRUE)
}
\arguments{
\item{Pdata}{A PacFIN dataset}

\item{keep_age_methods}{By default, methods "B","S" and "" are
retained.}

\item{minAge}{Minimum age to retain; default is 1.}

\item{maxAge}{Maximum age to retain; default (NULL) returns all ages
greater than \code{minAge}.}

\item{CLEAN}{Default is TRUE.  If FALSE, return the input data unchanged.}
}
\value{
The input data filtered for age methods and minimum and maximum ages
specified, with added column \code{age}.  A brief report on filtering is
output to the console.
}
\description{
\code{cleanAges} removes the samples with bad ages or agemethods.
}
\details{
\subsection{\code{\link{Workflow}}}{
In most cases, you should run \code{\link{cleanPacFIN}}
before \code{cleanAges}.  There may be cases where initial exploration of the
age data can be better accomplished without it.
}

Age methods "1" and "2" are equivalent to "B" and "S" respectively,
and are reset to those values.

No ages less than one occur in the commercial dataset.  Zero ages are
equivalent to NA.
}
\author{
Andi Stephens
}
\seealso{
\code{\link{cleanPacFIN}}, \code{\link{sink}}
}

