% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanColumns.R
\name{cleanColumns.bds}
\alias{cleanColumns.bds}
\title{Clean columns of PacFIN biological data}
\usage{
cleanColumns.bds(data)
}
\arguments{
\item{data}{A data frame with named columns.}
}
\value{
A data frame with fewer columns and some column names changed.
}
\description{
Many column names are changed in the biological data to match what was
used in legacy code. As time goes on the code will be reworked to use the
new name coming from PacFIN. Columns not used in the code are removed.
}
