% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanColumns.R
\name{cleanColumns.catch}
\alias{cleanColumns.catch}
\title{Clean columns of PacFIN catch data}
\usage{
cleanColumns.catch(data)
}
\arguments{
\item{data}{A data frame with named columns.}
}
\value{
A data frame with fewer columns.
}
\description{
Clean columns of a data frame from PacFIN that contains catch data.
The original column names are retained as there was no legacy code
for catch data. Many columns are removed and only those based on
the call to \link[dplyr:select]{dplyr::select} are kept.
}
