% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanColumns.R
\name{cleanColumns}
\alias{cleanColumns}
\title{Change Column Names to Vdrfd}
\usage{
cleanColumns(data)
}
\arguments{
\item{data}{A data frame with named columns.}
}
\value{
A data frame with fewer columns and some names changed.
}
\description{
Clean columns of a data frame from PacFIN.
All description columns are removed.
Should you want anything different, please feel free to
post an issue on github, email the package maintainer, or submit a
pull request.
}
\author{
Kelli Faye Johnson
}
