% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPacFIN.R
\name{cleanPacFIN}
\alias{cleanPacFIN}
\title{Filter PacFIN samples.}
\usage{
cleanPacFIN(Pdata, only_USINPFC = FALSE, keep_INPFC = NULL,
  remove_INPFC = NULL, badRecords = NULL, keep_sample_type = c("", "M"),
  keep_sample_method = "R", keep_missing_lengths = FALSE, keep_CA = TRUE,
  CLEAN = TRUE)
}
\arguments{
\item{Pdata}{a PacFIN dataset}

\item{only_USINPFC}{a logical value. FALSE by default.}

\item{keep_INPFC}{a set of INPFC areas. NULL by default.}

\item{remove_INPFC}{a set of INPFC areas.  NULL by default.}

\item{badRecords}{a set of sample identifiers. NULL by default.}

\item{keep_sample_type}{a set of sample types to retain.  Default = c("", "M")}

\item{keep_sample_method}{a set of sample methods to retain.  Default = "R"}

\item{keep_missing_lengths}{a logical value. FALSE by default.}

\item{keep_CA}{default TRUE.  CA data often have no sample type or method, or INPFC area.}

\item{CLEAN}{a logical value.  Default is TRUE.  If FALSE, return the original data unchanged,
but report what would have been removed.}
}
\value{
The input data filtered for desired areas and record types
specified, with added columns

\tabular{ll}{ 
  fishyr \tab initialized from SAMPLE_YEAR\cr
  fleet \tab initialized to 1\cr
  fishery \tab initialized to 1\cr
  season \tab initialized to 1.  Change using \code{\link{getSeason}}\cr
  state \tab initialized from SOURCE_AGID.  Change using \code{\link{getState}}\cr
  lengthcm \tab floored cm from FORK_LENGTH when available, otherwise FISH_LENGTH\cr
  geargroup \tab the gear group associated with each GRID, from http://pacfin.psmfs.org/pacfin_pub/data_rpts_pub/code_lists/gr.txt
}
}
\description{
\code{cleanPacFIN} filters out unsuitable samples from data, and converts
fish lengths to cm. The original fields in the returned data are left untouched,
with the exception of SEX, which is modified so that unidentified fish are labeled
"U".

\subsection{\code{\link{Workflow}}}{
If there are CalCOM samples to be integrated with PacFIN data, run \code{combineCalCOM}
first, otherwise run to \code{cleanPacFIN} as the first function in the workflow.
}
}
\details{
\subsection{\strong{INPFC Area specification}}{

The US INPFC areas are 
   c("VUS","CL","VN","COL","NC","SC","EU","CALCOM","CP","EK","MT","PS ")
   
"CalCOM" is included because the combineCalCOM function
sets it, since CalCOM doesn't seem to record INPFC areas.


If \code{only_USINPFC} is TRUE, then only samples from the US INPFC areas will be retained.

If a set of INPFC areas are specified in \code{keep_INPFC}, then only samples from 
those areas will be retained.

If \code{remove_INPFC} specifies a set of INPFC areas, samples from those areas
will be discarded.
}

\subsection{\strong{Sample types and methods}}{

SAMPLE_TYPEs may be (M=Market, R=Research, S=Special request, C=Commercial on-board).
Only samples of type M are generally used.

SAMPLE_METHODs may be (R=Random, S=Stratified, N=Systematic, P=Purposive, X=Special).
Only samples collected in random sampling are generally used.
}

\subsection{\strong{Furthermore}}{

The values created as new columns are for use by other functions in this package.
In particular, \code{fishyr} and \code{season} are useful if there are multiple 
seasons (e.g., winter and summer, as in the petrale sole assessment), and the 
year is adjusted so that "winter" occurs in one year, rather than across two.

The \code{fleet}, \code{fishery}, and \code{state} columns are meant for use in
stratifying the data according to the particulars of an assessment.

The \code{sink} command can be used to save the filtering report to a 
file, in addition to printing it to the console.
}
}
\seealso{
\code{\link{cleanAges}}, \code{\link{getState}}, \code{\link{getSeason}}, 
\code{\link{sink}}
}
\author{
Andi Stephens
}
