% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codify_age_method.R
\name{codify_age_method}
\alias{codify_age_method}
\title{Code ageing methods to standardized character strings}
\usage{
codify_age_method(x)
}
\arguments{
\item{x}{A vector that can be a mix of integers and characters.}
}
\value{
A vector of characters.
}
\description{
Code ageing error methods to a set of standard character strings because
some agencies use different codes for the same method. For example, break
and burn can be "B", "BB", or 1.
}
\details{
\subsection{Washington Department of Fish and Wildlife}{
\itemize{
\item B-break and burn;
\item E-enhanced break and burn;
\item L-length;
\item M-more than one age method was used, e.g., "B" and "S";
\item N-not aged;
\item O-optical scanner;
\item X-sectioning
}
}

\subsection{Oregon Department of Fish and Wildlife}{
\itemize{
\item 1-break and burn;
\item 2-surface;
\item 3-scales;
\item 4-thin section;
\item 5-optical scanner;
\item 6-length;
\item 9-unable
}
}

\subsection{California Department of Fish and Wildlife}{
\itemize{
\item B-break and burn;
\item S-surface;
\item T-thin section
}
}
}
\examples{
codify_age_method(c("B", "1", "AA", NA))
}
\author{
Kelli F. Johnson
}
