% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comps_wide.R
\name{comps_wide}
\alias{comps_wide}
\title{Convert long data frame with composition data to wide data frame}
\usage{
comps_wide(
  data,
  breaks,
  col_bins = "Age",
  col_proportions = "lf",
  includeplusgroup = TRUE
)
}
\arguments{
\item{data}{A data frame with named columns.}

\item{breaks}{A vector of cut points to be passed to \link{cut}.}

\item{col_bins}{The column name that pertains the bin or group you
wish to transform from long to wide, i.e., every unique value in this
column will be a new column in the wide format.}

\item{col_proportions}{The column name that pertains to the proportion
data in the data frame. This column will be used as information for each
new column in the wide format, i.e., summed across created categories.}

\item{includeplusgroup}{A logical defining if you want all values
larger than the largest \code{break} to be included in the last bin.
For example, if the largest value in \code{breaks} is 15 and you have
age-20 fish, then you should use \code{includeplusgroup = TRUE} if you
want to have data on this 20 year old fish. The default is to
include them.}
}
\value{
A data frame with initial columns for each stratification present
in the data, i.e., all columns not included in
\code{col_bins} and \code{col_proportions}. Followed by one column for each bin
present in the data that sums the proportional information for each
combination of stratification and bin. This data frame can often be used
as is in Stock Synthesis, though the stratification columns may need to
be altered such that sample sizes are correct or added if missing from
the stratification levels.
}
\description{
Convert a long data frame with one column for proportions, one
column for the bin each proportion belongs to, and stratification
columns that inform the model what time, sex, fleet, etc. that the
data come from.
}
\examples{
temp <- data.frame(
  state = rep(c("WA", "OR"), length.out = 30),
  year = rep(2010:2015, each = 5),
  Age = rep(1:15, 2),
  ap = rlnorm(n = 30))
comps <- comps_wide(temp, breaks = 3:8, col_proportions = "ap")
testthat::expect_equal(NCOL(comps), 8)
\dontrun{
print(comps)
}

}
\author{
Kelli F. Johnson
}
