% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSampleTable.R
\name{createSampleTable}
\alias{createSampleTable}
\title{Create table of samples by fleet and year to 
be included in US West Coast groundfish assessments}
\usage{
createSampleTable(
  Pdata,
  fname = NULL,
  strat = NULL,
  comps = "LEN",
  remove_yrs = NULL
)
}
\arguments{
\item{Pdata}{data frame}

\item{fname}{user specified file name (.csv) for the tables}

\item{strat}{calculate the sample sizes using the input columns ( c("fleet","usegear"))}

\item{comps}{specify whether to calculate the length or age samples}

\item{remove_yrs}{remove years of data for inclusion in the table.}
}
\description{
todo: document
}
\author{
Chantel Wetzel
}
