% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find.matching.rows.R
\name{find.matching.rows}
\alias{find.matching.rows}
\title{Function find.matching.rows}
\usage{
find.matching.rows(file, table, findex = 1, tindex = 1, tcol = 2,
  round. = T)
}
\description{
Function find.matching.rows
}
\details{
Using the primary keys in columns named 'findex' and 'tindex', finds the
matching values for 'file' in 'table' and returns 'table' column(s) 'tcol'.
Note that no test is made to determine if there are unmatched rows.
}
\author{
John R. Wallace (John.Wallace@noaa.gov), (revised) Andi Stephens, 2010.
}

