% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codelist.R
\name{get_codelist.GRID}
\alias{get_codelist.GRID}
\title{Get GRID options in PacFIN}
\usage{
get_codelist.GRID()
}
\value{
A data frame with the following six columns:
\itemize{
  \item{TYPE}{integer value of 1 (gear code), 2 (gear group), or 3 (all)}
  \item{GRID}{three letter grid code}
  \item{GROUP}{three letter gear code}
  \item{SHORT}{short description}
  \item{DESCRIPTION}{long description}
  \item{ENTERED}{date information was entered into table}
}
}
\description{
Get GRID options in PacFIN
}
