% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codelist.R
\name{get_codelist.INPFC}
\alias{get_codelist.INPFC}
\title{Get all INPFC Areas in PacFIN}
\usage{
get_codelist.INPFC()
}
\value{
A data frame with the following six columns:
\itemize{
\item{NAME}{a short-hand version of the area description}
\item{ARID}{mapping of sub area to areas}
\item{TYPE}{classification of tree structure, i.e., if it is a subarea}
\item{COUNCIL}{the management council that the area falls within}
\item{DESCRIPTION}{detailed description of the area}
\item{INPFC}{INPFC_AREA code used in PacFIN}
}
}
\description{
Get all INPFC Areas in PacFIN
}
