% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codelist.R
\name{get_codelist}
\alias{get_codelist}
\title{Get a list of available options, e.g., code_list, from PacFIN}
\usage{
get_codelist(x = c("GRID", "INPFC", "PORT"))
}
\arguments{
\item{x}{A character value specifying which code_list you want.
See the function call for the available lists or
\code{utils::methods("get_codelist")} for a list of all of the available
.S3methods.}
}
\value{
A data frame with the following potential columns, where the
column names are always in all caps:
\itemize{
\item{TYPE : }{integer value of 1 (gear code), 2 (gear group), or 3 (all)}
\item{GRID : }{three letter grid or gear type code}
\item{GROUP : }{three letter gear grouping code}
\item{ARID : }{mapping of sub area to areas}
\item{INPFC : }{International North Pacific Fisheries Commission (INPFC) Area code used in PacFIN}
\item{TYPE : }{classification of tree structure, i.e., if it is a subarea}
\item{COUNCIL : }{the management council that the area falls within}
\item{PCID : }{character port code}
\item{AGID : }{agency code, e.g., "W" for Washington}
\item{PORT : }{numberic port code}
\item{SHORT : }{short description}
\item{DESCRIPTION : }{long, detailed description}
}
}
\description{
Get a data frame with options available in PacFIN and their descriptions.
PacFIN documentation refers to these data frames as code_lists, thus the
function name \code{get_codelist} helps explain what we are getting.
}
\details{
S3:
This is an S3 class object, of which I am not very familiar with.
Thus, feel free to provide comments on the code or submit an
\href{https://github.com/pfmc-assessments/PacFIN.Utilities/issues}{issue} if
you have ideas or corrections.

source:
All PacFIN \href{https://pacfin.psmfc.org/pacfin_pub/codes.php}{code lists}
are publicly available. As time progresses, I hope to have all lists
available within R through \code{code_lists}. If you have time to add
any of this functionality, feel free to submit a pull request and
thank you in advance for collaborating.
}
\examples{
availablegrids <- get_codelist("GRID")
availableports <- get_codelist("PORT")

}
