% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeMethod.R
\name{getAgeMethod}
\alias{getAgeMethod}
\title{Get the ageing method used for the age that is agreed to be the best age}
\usage{
getAgeMethod(Pdata, verbose = TRUE)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fishieries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
A vector of characters.
}
\description{
Each age read will have an associated age method. This function will attempt
to find what ageing method was used for the age thought to represent the
best estimate of the age of the otolith.
}
\details{
When more than one age estimate matches what is thought to be the best age,
then the ageing methods for all ages that match the best age will be
returned as a single entry separated by "--". Most often, double reads are performed using the same method but there are cases, e.g., CARE exchanges or research questions on differences between methods, when a single otolith is read using multiple methods. For example, some petrale sole otoliths were read using both surface and break and burn reads.
}
\examples{
data <- data.frame(
  age1 = 1:10,
  age2 = 10,
  AGE_METHOD1 = "B",
  AGE_METHOD2 = "S",
  FINAL_FISH_AGE_IN_YEARS = 10
)
getAgeMethod(data)
# [1] "S"    "S"    "S"    "S"    "S"    "S"    "S"    "S"    "S"    "B--S"
}
\author{
Kelli F. Johnson
}
