% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArea.R
\name{getArea}
\alias{getArea}
\title{Get the general area that the information was taken from}
\usage{
getArea(Pdata, verbose = TRUE)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
A vector of characters with the same length as \code{NROW(Pdata)}. Rows of data
that should potentially be removed from your data are given a grouping like
\code{"Puget Sound"} in an attempt to indicate where the record came from in a
more transparent way than cryptic codes like \verb{4a}. \code{NA}s indicate records
that the package maintainers believe should be maintained in your data set.
}
\description{
Get the general area that the information (i.e., row of data) was taken from.
Thankfully, PacFIN contains records from all areas, but unfortunately, we do
not always want to use data from all of these areas in the stock assessment.
Thus, this function helps to identify which data should be included based on
standard filtering. We encourage users to continue to explore their data for
records that may be outside of their jurisdiction for further filtering.
}
\section{Rationale for excluded areas}{
\subsection{Washington Department of Fish and Wildlife (WDFW)}{

WDFW uses management regions, also known as Marine Fish-Shellfish Management
and Catch Reporting Areas, to define the area in which landings occurred and
sampled fish come from. These management regions are numeric and fall within
the range of \code{1:99}, though not all numeric values within this range are
actually used. WDFW uses their management regions to define the PSMFC area in
which landings occurred because PSMFC area is not a ``recorded'' value on
fish tickets. See Table 3 on page 15 of \href{https://wdfw.wa.gov/sites/default/files/publications/01754/wdfw01754.pdf}{Tsou et al. (2015)}
for a table that converts WDFW management region to PSMFC area.

A partial map of the WDFW management regions for the Puget Sound areas is
provided in the \href{https://wdfw.wa.gov/sites/default/files/publications/00927/wdfw00927.pdf}{Puget Sound Groundfish Management Plan}
(Palsson et al., 1998; see Figure 3 on page 26), and an arcgis layer of the
management regions within the Strait of Juan de Fuca and the Puget Sound was
posted by
\href{https://www.arcgis.com/apps/webappviewer/index.html?id=13a09a6daae7408383fe3617ced75259}{SITC}.
Both of these resources were helpful in understanding how WDFW management
regions 20-29 correspond to PSMFC area 4A and should be attributed to the
Puget Sound. Where, 4A corresponds to everything east of Cape Flattery (i.e.,
the most easterly point of Washington state).
}

\subsection{Foreign}{

Foreign samples, those coming from outside of the EEZ, could be in your data.
The PacFIN documentation for
\href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/ar_tree.txt}{areas}
identifies areas \code{ZY} or \code{ZZ} or \code{9[GMYZ]} as those outside of the EEZ.
Additionally, 3D and 3N make up VC, which is Canadian catches within the
Vancouver area.
}
}

\seealso{
\code{\link[=cleanPacFIN]{cleanPacFIN()}} uses \code{getArea()}.
}
