% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComps.R
\name{getcomps_long}
\alias{getcomps_long}
\title{Create a Long Database to Prepare Compositions}
\usage{
getcomps_long(
  data,
  towstrat,
  type,
  towid = "SAMPLE_NO",
  weightid = "Final_Sample_Size_L",
  dropmissing = TRUE
)
}
\arguments{
\item{data}{A data frame with columns defined by the following
arguments: towstrat, type, towid, and weightid}

\item{towstrat}{A vector of character values providing the column
names for which you want compositions for.}

\item{type}{A character value specifying which category to
summarize by, i.e., \code{"length"} or \code{"Age"}.}

\item{towid}{A vector of character values providing the column
names that generate a unique id for each sample.}

\item{weightid}{A character value giving the column name that
holds the value to be summed for each type and strata.}

\item{dropmissing}{A logical value supplied to the
\code{drop} argument in \code{stats::aggregate}
that specifies whether or not to keep all levels in the data
even if there are no values to report for summaries.}
}
\value{
A data frame in long form with a weight for each
category included in the lengths or ages of interest by
stratification. Stratifications are normally year and fishery
and sex will always be included. If no sex is provided then it
is assumed all are unsexed and will be returned as such.
}
\description{
Create a Long Database to Prepare Compositions
}
\author{
Kelli F. Johnson
}
