% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDB.R
\name{getDB}
\alias{getDB}
\title{Get a Connection to a Database and Return a Query}
\usage{
getDB(
  sql,
  datasourcename = "PacFIN",
  username = getUserName(datasourcename),
  password,
  asis = FALSE
)
}
\arguments{
\item{sql}{A valid SQL statement.}

\item{datasourcename}{The sql data base name that is stored on your
computer. The default is \code{"PacFIN"} which will pull data from
the Pacific Fisheries Information Network.}

\item{password}{Your password for the \code{datasourcename}.}

\item{asis}{A logical, specifying whether or not to convert columns,
as in \code{\link[utils]{read.table}}.}
}
\value{
A data frame resulting from querying the database or an
error message because the main call is wrapped in
\code{\link[base]{tryCatch}}.
}
\description{
Get a connection established to a database and return the information desired
in the specific query.
}
\author{
John R. Wallace, Kelli Faye Johnson
}
