% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGearGroups.R
\name{getGearGroups}
\alias{getGearGroups}
\title{Create column for gears according to PacFIN gears}
\usage{
getGearGroups(Pdata, gears = NULL)
}
\arguments{
\item{Pdata}{A PacFIN \code{data.frame}}

\item{gears}{A data.frame or \code{file.path} to a \code{.csv} file
that contains GRIDS according to the table taken from
[gear table](http://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/gr.txt).
Path name must be the full path or in your current working directory
and end in \code{.csv}. If \code{gears = NULL} then the function assumes
that the appropriate table is downloaded and available in your working
directory saved as \code{'GearTable.csv'}.}
}
\value{
A modified \code{data.frame} where an additional column labeled
  \code{geargroups} is added to \code{Pdata}. No original columns are modified
  in the process.
}
\description{
Data from the PacFIN
[gear table](http://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/gr.txt)
is used to create a column in \code{Pdata} labeled \code{geargroup}, where
\code{Pdata$GRID} is recoded to \code{geargroup} according to the gear table.
}
\author{
Andi Stephens
}

