% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMed.R
\name{getMed}
\alias{getMed}
\title{Calculate Category-Specific Medians}
\usage{
getMed(x, ...)
}
\arguments{
\item{x}{A vector of values you want split by categories listed
in \code{...}. Number of values in \code{x} must be the same
length as each vector supplied in \code{...}.}

\item{...}{Categories that you want to use to split the data.
Any number can be supplied.}
}
\value{
A vector of medians with one entry for each original
entry because the results are duplicated for each entry that
went into the calculation.
}
\description{
Compute medians by categories with \code{NA} values omitted.
This eliminates the need to match results back to the 
original data set using merge or indexing functions.
}
\examples{
data <- data.frame(
  "cat1" = rep(c("yes", "no"), each = 10),
  "cat2" = rep(1:5, each = 4),
  "val" = rnorm(20))
PacFIN.Utilities:::getMed(x = data[, "val"], 
  data[, "cat1"], data[, "cat2"])

}
\author{
Kelli Faye Johnson
}
