% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPacFINgears.R
\name{getPacFINgears}
\alias{getPacFINgears}
\title{Read PacFIN Gear Type List from URL}
\usage{
getPacFINgears()
}
\value{
A data frame of gear types with descriptions from PacFIN and 
the gear group for each gear type.
}
\description{
Download the list of PacFIN gears from
https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/gr.txt
}
\examples{
gear <- getPacFINgears()
rm(gear)

}
\author{
Kelli Faye Johnson
}
