% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeason.R
\name{getSeason}
\alias{getSeason}
\title{Add a column to \code{Pdata} for season.}
\usage{
getSeason(
  Pdata,
  season_type = -1,
  yearUp = NULL,
  yearDown = NULL,
  plotResults = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fishieries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{season_type}{Specify a \code{numeric} value for season type.
If negative then all months will be assigned to season \code{1}.
If \code{0} then seasons will be assigned from \code{Pdata$SAMPLE_MONTH},
where each month is a unique season.
If \code{1} then seasons are assigned according to methods used for Petrale,
where winter months (\code{c(11:12, 1:2)}) are season \code{1} and
the remaining months (summer) are assigned to season \code{2}.
Please contact the package maintainer should you wish to include an
additional seasonal scheme.}

\item{yearUp}{Used to provide a list of months (i.e., \code{1:12})
for which to adjust the year (\code{Pdata$fishyr}) up. For example,
if winter months belong to the following year then use \code{yearUp = 11:12}.}

\item{yearDown}{Used to provide a list of months (i.e., \code{1:12})
for which to adjust the year (\code{Pdata$fishyr}) down. For example,
if winter months belong to the previous year then use \code{yearUp = 1:2}.}

\item{plotResults}{A logical value specifying if plots should or should not
be created and shown in the console.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
An additional column \code{season} is added to \code{Pdata}.
No columns are modified.
}
\description{
Several seasonal schemes are available, including Petrale seasons
(1 = winter months, 2 else).
Most assessments won't require \code{getSeason} because it is included
in \link{cleanPacFIN}. If a specialized season structure is required,
\code{getSeason} should be run immediately after \link{cleanPacFIN}.
}
\examples{
test <- getSeason(
  data.frame(SAMPLE_MONTH = 1:12, fishyr = rep(1:2, each = 6)),
  verbose = TRUE
)
testthat::expect_true(all(test[, "season"] == 1))
test <- getSeason(Pdata = test, season_type = 1, yearUp = 12)
testthat::expect_equivalent(test[test[, "fishyr"] == 3, "season"], 1)

}
\author{
Andi Stephens
}
