% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComps.R
\name{getunsexedsamps}
\alias{getunsexedsamps}
\title{Return Sample IDs That Did Not Sex Samples}
\usage{
getunsexedsamps(identifier, sex, good = "U")
}
\arguments{
\item{identifier}{Unique IDs for the samples}

\item{sex}{A vector of the same length as \code{identifier} providing
the sex of the sampled fish.}

\item{good}{A character value supplying the \code{sex} value you
want to keep.}
}
\value{
A vector of identifiers that only had the sex given in
the \code{good} argument. The returned vector will be of the same
length as the supplied vectors.
}
\description{
Identifiers that have female and male samples will be returned 
as NA and only identifiers that had unsexed fish will be provided.
}
\author{
Kelli Faye Johnson
}
