% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getweight.R
\name{getweight}
\alias{getweight}
\title{Calculate Weight From Parameters}
\usage{
getweight(length, sex, pars)
}
\arguments{
\item{length}{A vector of fish lengths in mm.}

\item{sex}{A vector of sexes for each entry in \code{length}.
Entries must be one of the following: 'U', 'F', 'M', or 'H'.}

\item{pars}{A data frame of parameters for the weight-length
relationship as determined from empirical data. The data frame
must have columns of \code{'A'} and \code{'B'}, as well as 
rows named \code{'U'}, at a minimum, \code{'F'}, and \code{'M'}.
The output from \code{\link{getWLpars}} is formatted correctly as is.}
}
\value{
A vector of weights determined from lengths and weight-length
parameters input to the function. 
Weights are in pounds.
}
\description{
Estimate fish weights from sex-specific, or not, parameters
estimated, or formated the same way, as that returned from
\code{\link{getWLpars}}.
}
\author{
Kelli Faye Johnson
}
