% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAge.R
\name{grep_final_age}
\alias{grep_final_age}
\title{Find the column with the best age information}
\usage{
grep_final_age(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A single character object.
}
\description{
Search a vector of strings for \code{"FINAL_FISH_AGE_IN_YEARS"} or
\code{"FISH_AGE_YEARS_FINAL"}.
}
\author{
Kelli F. Johnson
}
