% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComps.R
\name{lenique}
\alias{lenique}
\title{Number of Unique Entries
A helper function that returns the number of unique
entries in a vector. Usefull for apply functions.}
\usage{
lenique(x)
}
\arguments{
\item{x}{A vector of character or numeric values.}
}
\value{
An integer value specifying the number of unique entries.
}
\description{
Number of Unique Entries
A helper function that returns the number of unique
entries in a vector. Usefull for apply functions.
}
\author{
Kelli Faye Johnson
}
