% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_INPFC.R
\name{ls_INPFC}
\alias{ls_INPFC}
\title{List International North Pacific Management Council (INPFC) Areas}
\usage{
ls_INPFC(area = c("ALL", "US"))
}
\arguments{
\item{area}{A valid group that you want to look up. Available options for this
argument are listed in the function call. The function is not vectorized,
i.e., only allows a single input value.}
}
\value{
A vector of character values within the group specified.
}
\description{
List (ls) INPFC areas of interest.
}
\examples{
ls_INPFC(area = "US")

}
