% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PacFIN.Utilities.R
\docType{package}
\name{PacFIN.Utilities}
\alias{PacFIN.Utilities-package}
\alias{PacFIN.Utilities}
\title{PacFIN.Utilities: Functions for working up PacFIN data}
\description{
The PacFIN.Utilities package provides functions for filtering, summarizing, expanding, and compiling
composition data, and writing out the final products:
\itemize{
\item length-
\item age- and
\item age-at-length compositions.
}
}
\details{
Many of the functions described below write summary output to the console, and several create plots for
visualizing data at various steps in processing.  The information written to the console
can be captured using the \link{sink} function to encapsulate your workflow:

\code{sink("myfilename.txt", split=TRUE)}

... do work ...

\code{sink()}

The second call closes the file.
}
\section{Filtering functions}{

\link{cleanPacFIN} Filters raw PacFIN data
}

\section{Summary functions}{

\link{plotRawData} Plots commonly viewed raw PacFIN data

\link{plotCleaned} Plots the product of \code{cleanPacFIN}

In addition, the \link{getExpansion_1} function has a "plot" argument to create
plots documenting the expansions it creates.  It also writes summary output to the
console.
}

\section{Optional functions}{

\link{getSeason} For treating data from seasonal fisheries, such as Petrale.
}

\section{Expansion functions}{

\link{getExpansion_1} creates the values for the sample to tow expansion.

\link{getExpansion_2} computes the expansion values from the tow upwards to a
user-specified stratification
}

\section{Expansion caveats}{

There is one manual step in the workflow.
After running the expansion functions, data columns Expansion_Factor_1 and
Expansion_Factor_2 are available to use in manually setting the Final_Expansion_Factor.
\itemize{
\item Age data are expanded separately from lengths
\item WA fish are generally only expanded using Expansion_Factor_2.
\item Other expansions are the product of Expansion_Factor_1 * Expansion_Factor_2
\item For age-at-length comps, set Final_Expansion_Factor to 1.  Each fish represents only itself.
}
}

\section{Composition functions}{

\link{getComps} uses the column specified in \code{weightid} and the user-specified
stratification to create comps.

\link{doSexRatio} separates the unsexed fish into males and females

\link{writeComps} formats the composition data for SS3 and writes it to a file.
}

\section{Infrequently used functions}{

You can run these manually to reset the initial values.

\link{getState} Run by \link{cleanPacFIN} to initialized the \code{state} column.

\link{getGearGroup} Run by \link{cleanPacFIN} to initialize \code{geargroups}.

\link{capValues} Used to limit the maximum value in a vector.
}

\author{
\strong{Maintainer}: Kelli F. Johnson \email{kelli.johnson@noaa.gov} (\href{https://orcid.org/0000-0002-5149-451X}{ORCID})

Authors:
\itemize{
  \item Chantel R. Wetzel \email{chantel.wetzel@noaa.gov} (\href{https://orcid.org/0000-0002-7573-8240}{ORCID})
}

Other contributors:
\itemize{
  \item Kathryn L. Doering \email{kathryn.doering@noaa.gov} (\href{https://orcid.org/0000-0002-0396-7044}{ORCID}) [contributor]
  \item Brian J. Langseth \email{brian.langseth@noaa.gov} (\href{https://orcid.org/0000-0002-9901-6146}{ORCID}) [contributor]
  \item Andi Stephens \email{andi.stephens@noaa.gov} [contributor]
  \item Ian G. Taylor \email{ian.taylor@noaa.gov} (\href{https://orcid.org/0000-0002-4232-5669}{ORCID}) [contributor]
  \item John R. Wallace \email{john.wallace@noaa.gov} (\href{https://orcid.org/0000-0002-2333-1262}{ORCID}) [contributor]
}

}
