% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paste.col.R
\name{paste.col}
\alias{paste.col}
\title{Utility function.}
\usage{
paste.col(x)
}
\description{
Used by \code{\link{find.matching.rows}}.
}
\details{
Converts each row in the input dataframe to a string,
"pasting" the columns together.
}

