% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCleaned.R
\name{plotCleaned}
\alias{plotCleaned}
\title{Plot cleaned fishery age- or length-composition data.}
\usage{
plotCleaned(Pdata, savedir = getwd())
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fishieries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}
}
\value{
The following figures are saved to \code{savedir}:
\itemize{
\item PacFIN_comp_Nbystate.png
\item PacFIN_comp_distributions.png
\item PacFIN_comp_NbyGRID.png
\item PacFIN_comp_depth.png
\item PacFIN_comp_geargroup.png
\item PacFIN_comp_INPFC.png
\item PacFIN_comp_lengthvage.png
}
}
\description{
Save several figures to your computer that summarize the composition
samples that are available for the given species. None of the figures
are meant to be included in the stock assessment document, but rather,
are for exploring the data.
}
\seealso{
This function is called by \link{cleanPacFIN} and heavily
relies on \link{getGearGroup} to create gear categories.
}
\author{
Andi Stephens, Kelli F. Johnson
}
