% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRawData.R
\name{plotRawData}
\alias{plotRawData}
\title{Diagnostic plots and summaries for a raw PacFIN dataset pre-filtering.}
\usage{
plotRawData(rawData, fname = NULL)
}
\arguments{
\item{fname}{a filename}

\item{Rawdata}{an unfiltered PacFIN dataset}
}
\description{
\code{plotRawData} creates a set of diagnostic plots and summaries, writing 
pdfs and a text file in addition to plotting onscreen and console.
}
\details{
\subsection{Workflow}{
Run \code{plotRawData} to visualize and summarize the PacFIN data prior to
running \code{\link{cleanPacFIN}}.
}


Will create a filename from the species ID if one is not provided.
}

