% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStrat.R
\name{plotStrat}
\alias{plotStrat}
\title{Plot Length Distributions with Aged and All Fish}
\usage{
plotStrat(
  data,
  dir = getwd(),
  npages = 5,
  col.fleet = "fleet",
  col.area = "state",
  col.age = "Age",
  col.length = "lengthcm",
  height = 10,
  width = 10
)
}
\arguments{
\item{data}{A data frame}

\item{dir}{The directory you want to print the plots}

\item{npages}{The number of pages you want the plots split over.}

\item{col.fleet}{Fleet column name.}

\item{col.area}{Area column name.}

\item{col.age}{Age column name.}

\item{col.length}{Length column name.}

\item{height}{Figure height.}

\item{width}{Figure width.}
}
\description{
todo: document
}
