% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullBDS.PacFIN.R
\name{PullBDS.PacFIN}
\alias{PullBDS.PacFIN}
\title{Pull biological data from PacFIN database}
\usage{
PullBDS.PacFIN(
  pacfin_species_code,
  username = getUserName("PacFIN"),
  password,
  savedir = getwd()
)
}
\arguments{
\item{pacfin_species_code}{A character string for the PacFIN
species code that you desire to pull data for. Known also as
\code{"SPID"} in the PacFIN database if using legacy sql scripts.
An example for sablefish would be
\code{pacfin_species_code = "SABL"}.}

\item{username}{The username you would like to use to access data from
\code{datasourcename}. The default is to generate \code{username}
from your computer name.}

\item{password}{Your password for the \code{datasourcename}.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}
}
\value{
RData frame is saved to the disk and the pulled data
are returned as a data frame.
The data frame is fairly raw in its form with only one column
added, \code{all_cluster_sum}, which is the weight of all clusters sampled
for a given sample number. Typically, \link{cleanPacFIN} will be used
on the data frame after its retrieval.
}
\description{
Pull biological data from the Comprehensive Biological Data Table
in the PacFIN database.
}
\examples{
\dontrun{
# Enter password in place of xxx below
mypw <- "xxx"
pd <- PullBDS.PacFIN(pacfin_species_code = "POP", password = mypw)
}

}
\seealso{
\link{cleanPacFIN} to manipulate and subset the returned object.
}
\author{
John R. Wallace, Kelli Faye Johnson
}
