% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullNominal.PacFIN.R
\name{PullNominal.PacFIN}
\alias{PullNominal.PacFIN}
\title{Pull the species information table and return nominal code(s)}
\usage{
PullNominal.PacFIN(
  pacfin_species_code,
  username = getUserName("PacFIN"),
  password = ask_password()
)
}
\arguments{
\item{pacfin_species_code}{A vector of strings specifying the PacFIN species
code(s) you are interested in. This has sometimes been referred to as
\code{"SPID"} in legacy sql scripts. An example for sablefish would be
\code{pacfin_species_code = "SABL"}. Lists of species codes in \href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/sp_tree.txt}{hierarchical order},
by
\href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/agency_spcodes.txt}{organization},
and \href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/agency_spcodes_spid.txt}{alphabetically organized}
can be found on the \href{https://pacfin.psmfc.org}{PacFIN website}. Often you
will want to include nominal species categories. Where, nominal (i.e.,
\href{https://www.google.com/search?q=NOMINAL&rlz=1C1GCEJ_enUS866US866&sourceid=chrome&ie=UTF-8}{existing in name only})
means information for a given species that is "derived" from non-species
specific information, e.g., species complexes that are split out by species
compositions like "nominal aurora rockfish" which would be \code{ARR1}. For some
functions, these nominal categories can automatically be added, see the
argument \code{addnominal}.}

\item{username}{Most often, this is a string containing your username for the
database of interest. You can use \code{\link[=getUserName]{getUserName()}} if you prefer to not
enter this argument and assume the default search and/or rules for finding
your username will work. This is the default behavior if you leave
\code{username} as a missing argument, i.e., \code{username <- getUserName(database = database)}. Sometimes this search will fail because of legacy rules, which
are unknown to the development team, that were used to create your
username. Please email the maintainer of this package if you need more
functionality here.}

\item{password}{Most often, this is a string containing your password for
the database of interest. You can use the function \code{\link[=ask_password]{ask_password()}} if you
would prefer to be prompted for your password. Please do not share this
password with anyone or push code to a repository that has your password
saved in it.}
}
\value{
A unique vector of character entries, where entries are valid
\code{PACFIN_SPECIES_CODE}s. The length of the returned vector may not be the same
length as the supplied \code{pacfin_species_code} argument because duplicates are
removed. \code{NA} is returned if no values are found.
}
\description{
Pull the species information table and attempt to determine which
nominal codes pertain to the desired species codes.
}
