% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.area.R
\name{sql.area}
\alias{sql.area}
\title{Write SQL Text for Area}
\usage{
sql.area()
}
\value{
A single character string formatted as an sql call.
}
\description{
Write SQL text as a single character string that will result in
getting the area data from the PacFIN database.
}
\author{
John R. Wallace, Kelli F. Johnson
}
