% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.catch.R
\name{sql.catch}
\alias{sql.catch}
\title{Write SQL Text for Catches}
\usage{
sql.catch(pacfin_species_code)
}
\arguments{
\item{pacfin_species_code}{A character string for the PacFIN
species code that you desire to pull data for. Known also as
\code{"SPID"} in the PacFIN database if using legacy sql scripts.
An example for sablefish would be
\code{pacfin_species_code = "SABL"}.}
}
\value{
A single character string formatted as an sql call.
}
\description{
Write SQL text as a single character string that will result in
getting the catch data from the PacFIN database.
}
\author{
Kelli Faye Johnson
}
