% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-.R
\name{sql}
\alias{sql}
\alias{sql_area}
\alias{sql_bds}
\alias{sql_catch}
\alias{sql_species}
\title{Write SQL text}
\usage{
sql_area()

sql_bds(pacfin_species_code)

sql_catch(pacfin_species_code, council_code = "P")

sql_species()
}
\arguments{
\item{pacfin_species_code}{A vector of strings specifying the PacFIN species
code(s) you are interested in. This has sometimes been referred to as
\code{"SPID"} in legacy sql scripts. An example for sablefish would be
\code{pacfin_species_code = "SABL"}. Lists of species codes in \href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/sp_tree.txt}{hierarchical order},
by
\href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/agency_spcodes.txt}{organization},
and \href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/agency_spcodes_spid.txt}{alphabetically organized}
can be found on the \href{https://pacfin.psmfc.org}{PacFIN website}. Often you
will want to include nominal species categories. Where, nominal (i.e.,
\href{https://www.google.com/search?q=NOMINAL&rlz=1C1GCEJ_enUS866US866&sourceid=chrome&ie=UTF-8}{existing in name only})
means information for a given species that is "derived" from non-species
specific information, e.g., species complexes that are split out by species
compositions like "nominal aurora rockfish" which would be \code{ARR1}. For some
functions, these nominal categories can automatically be added, see the
argument \code{addnominal}.}

\item{council_code}{A vector of character strings specifying the council code
that you wish to retain data for. The default \code{"P"} means data will only be
returned for the Pacific Fisheries Management Council. Other accepted
values are \code{"N"} or \code{"*"}.}
}
\value{
A single character string formatted as an sql call.
}
\description{
Write SQL text as a single character string that will result in getting the
relevant data from the PacFIN database.
}
\details{
\code{sql_area()} results in area data

\code{sql_bds()} results in biological data

\code{sql_bds()} results in catch data

\code{sql_species()} results in data frame of species names
}
\author{
Kelli F. Johnson and John R. Wallace
}
