% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.species.R
\name{sql.species}
\alias{sql.species}
\title{Write SQL Text for Species}
\usage{
sql.species()
}
\value{
A single character string formatted as an sql call.
}
\description{
Write SQL text as a single character string that will result in
getting the species data from the PacFIN database.
}
\author{
Kelli F. Johnson
}
