% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stratify.R
\name{Stratify}
\alias{Stratify}
\title{Create a vector of stratifications}
\usage{
Stratify(inVector = NULL, splits = NULL, names = NULL, numeric = F)
}
\arguments{
\item{inVector}{Input vector of values.  May be numeric or character.}

\item{splits}{Vector of values to use to stratify the input vector.}

\item{names}{Vector the length of \code{splits} naming the strata.}

\item{numeric}{A logical value, where if \code{TRUE} then
  \code{findInterval} will be used and stratum 0 will be assigned to values
  smaller than the first value in splits.}
}
\value{
Todo.
}
\description{
\code{Stratify} takes an input vector and list of values used to designate
the values returned in the strats vector.
\subsection{Workflow}{
\code{Stratify} is intended to add levels for stratification before
running \code{\link{getComps}}.
}
}
\details{
Any values not given in \code{splits} will be assigned to stratum 0.
This function can be used to designate fleet or for stratification
based on depth or INPFC area.
}
\examples{
Pdata$use_depth = Stratify( Pdata$DEPTH_AVG, splits=c(0, 50, 100, 250, 500), numeric=T)

Pdata$my_fleets = Stratify(( Pdata$GEAR,
           splits=list("GFS", c("MDT","TB"), "TR", c("LGL","FTS")),
           names=list(“G1”,”Troll”,”Trawl”,”G3”) )
}
\author{
Andi Stephens
}

