% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAgeMethod.R
\name{summaryAgeMethod}
\alias{summaryAgeMethod}
\title{The number of samples per combination of ageing methods}
\usage{
summaryAgeMethod(Pdata, verbose = FALSE)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
A data frame of counts (\code{n}) for each age method with the following columns:
\itemize{
\item \verb{AGE_METHOD[0-9]*}: where the number at the end of the column indicates
which age reader, e.g., first, second, third, etc., the method comes from.
This column contains a letter or number indicating the ageing method used.
There will be one column per number of reads available, i.e., if at least
one fish was read by four reader, then there will be four columns starting
with \code{AGE_METHOD}.
\item \verb{Age method for best age}: provides data on all methods used across age
readers, for example if both break and burn and surface read information is
available for a given fish then this column will have \code{B--S} as an entry.
This information is important because sometimes the best age is based off
multiple methods and it is difficult to know which ageing error to assign
the entry to.
\item \code{n}: the number of samples available for the given methods listed in that
row.
}

If \code{verbose}, then text explaining the summary is also printed to the screen
prior to returning the table.
}
\description{
Uses \code{\link[dplyr:count]{dplyr::count()}} to summarize the number of samples across the various
ageing methods (i.e., \verb{AGE_METHOD[0-9]+}) that were used in the passed data
frame. Sometimes the ageing method will differ between double reads or
between years.
}
\author{
Kelli F. Johnson
}
