% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeComps.R
\name{writeComps}
\alias{writeComps}
\title{Write out composition data formatted for Stock Synthesis}
\usage{
writeComps(
  inComps,
  fname = NULL,
  abins = NULL,
  lbins = NULL,
  maxAge = Inf,
  month = 7,
  partition = 2,
  ageErr = 0,
  dummybins = FALSE,
  sum1 = FALSE,
  digits = 4,
  overwrite = lifecycle::deprecated(),
  verbose = FALSE
)
}
\arguments{
\item{inComps}{A dataframe generated using \code{\link[=getComps]{getComps()}}.}

\item{fname}{A filename with the appropriate extension, used to save the
function output to the disk. For example, \code{LenComps.csv} or
\code{file.path(getwd(), "SampleSize.csv")}. Full, relative, or simple paths are
allowed because the argument is used as is, i.e., not redirected to a
directory different than \code{\link[=getwd]{getwd()}}. Note that various functions within
PacFIN.Utilities have different default values for this input argument. If
\code{NULL} in \code{writeComps}, then the resulting file name will be based on what
type of composition data is being generated, i.e., \code{PacFIN_lengths.out},
\code{PacFIN_ages.out}, or \code{PacFIN_CAAL.out} for length, age, or conditional
length-at-age data, respectively.}

\item{abins, lbins}{The binning structure to use for ages and lengths. For
both arguments, the default is \code{NULL} which leads to the natural bins of
the data being used, i.e., no additional binning is performed.}

\item{maxAge}{A numeric value specifying the maximum age of fish that
should be included in the composition data, unless \code{dummybins = TRUE},
then those fish will be included in a plus group which you can investigate
later. Note that \code{maxAge} is only used if \code{abins = NULL}, otherwise
fish are binned according to user specified bins irregardless of \code{maxAge}.}

\item{month}{Month for all observations. Defaults to 7. If input has multiple
seasons, this must be a vector of equal length to the maximum seasons where
the order of months in the vector will be assigned to season in ascending
order. For example, if there  are two seasons and the month = c(1, 7)
season 1 will be assigned to month 1 and season 2 to month 7.}

\item{partition}{Used by Stock Synthesis for length- or age-composition data
where 0 = retained + discarded, 1 = discarded, and 2 = retained fish. The
default is to assume that these fish are retained only. The default was
changed in 2020 from a value of 0, and code should be updated accordingly
if you really want 0.}

\item{ageErr}{Defaults to 1 to use the first ageing error matrix specified in
your Stock Synthesis model.}

\item{dummybins}{A logical value specifying whether data outside of the lower
and upper \code{abins} or \code{lbins} should be added to dummy bins, or be placed in
the specified bins. Default is \code{TRUE}.  Dummy bins are useful for
determining whether the current bin structure properly captures the data.}

\item{sum1}{A logical value specifying whether to rescale the compositions
to sum to 1 (they will also be rounded to 4 decimal places)}

\item{digits}{Integer indicating the number of decimal places to round
value to (value is passed to \code{\link[=round]{round()}}). NULL will skip rounding.}

\item{overwrite}{Deprecated as of version 0.2.10. The file will be
automatically overwritten if present.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
A \code{.csv} file is created and written to with composition data. And, a list
with all four types of composition data, i.e., different sex stratifications,
are returned as a list.
}
\description{
Write out composition data to a file, binning the data as specified to work
in a Stock Synthesis model.
}
\details{
The structure of the input dataframe determines whether the results
produce age-, length-, or conditional-age-at-length-
composition data.

Four sets of composition data are written to a single file specified
by \code{fname}. These are:
\itemize{
\item{females followed by males}
\item{male data only}
\item{female data only}
\item{males and females recombined as unsexed fish}
}

Composition data are raw weights rather than proportions. Stock Synthesis
internally converts these to proportions.  The raw weights should be
examined for anomalies.

To create proportions use \code{\link[=prop.table]{prop.table()}} on the columns
containing composition data.

The columns in the output preceding 'lengthcm' or 'age' are those that
were used in stratifying the data.

\subsection{Reality Checks}{
\itemize{
\item{Set \code{verbose = TRUE} to follow progress}
\item{Use \code{dummybins} to evaluate bin structure}
\item{Examine the raw-weight output for anomalies}
}}
}
\seealso{
\itemize{
\item \code{getComps()}
}
}
\author{
Andi Stephens, Chantel R. Wetzel, Kelli F. Johnson, Ian G. Taylor
}
