% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_representativeness_plot.R
\name{age_representativeness_plot}
\alias{age_representativeness_plot}
\title{Compare length histogram of aged samples to all lengthed samples}
\usage{
age_representativeness_plot(
  bio,
  xlim = c(0, max_break),
  ylim = c(0, 0.049),
  max_break = 155,
  file = NULL,
  plot_panels = c(10, 2),
  wait2plot = FALSE
)
}
\arguments{
\item{bio}{The dataframe containing the lengths and ages. Currently the
function works with data from the West Coast Groundfish Bottom Trawl
Survey extracted by the nwfscSurvey package as well as PacFIN data.
Must include either the columns Year, Length_cm, and Age, or,
for PacFIN data, SAMPLE_YEAR, FISH_LENGTH, and FISH_AGE_YEARS_FINAL.
Additional columns can be present.}

\item{xlim}{The x limits for the histogram}

\item{ylim}{The y limits for the histogram}

\item{max_break}{The maximum length in cm (If PacFIN data is provided,
FISH_LENGTH is automatically converted from mm to cm).}

\item{file}{The relative or absolute path and name of the file to write to,
as a .png file.}

\item{plot_panels}{A vector of two integers that determine the matrix
dimensions of the plots panels (row x columns), i.e., used in \code{par()}
if \code{!is.null(file)}.}

\item{wait2plot}{A logical specifying if you want R to ask you to confirm
page changes between plots. The default is \code{FALSE} and it will also
be \code{FALSE} if plots are saved to the disk.}
}
\value{
Several plots, potentially, are printed to the screen or saved to the disk
}
\description{
Compare length histogram of aged samples to all lengthed samples
}
\details{
Output on figures include K-S p-value, which is the p-value output from
\code{\link{ks.test}} and is colored red if <.05 and green if >=.05, and
\code{expression(hat(b))}, which is the bhattacharyya coefficient calculated
as \code{sum(sqrt(pi1*pi2))} where pi1 and pi2 are the proportion of data in
bin i for each of the two distributions.
}
\author{
Ian G. Taylor, Kathryn L. Doering, Brian Langseth, Kelli F. Johnson
}
