% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_representativeness_plot.R
\name{changecol_pacfin}
\alias{changecol_pacfin}
\title{Change Column Names of PacFIN Data To Match Survey}
\usage{
changecol_pacfin(data, verbose = FALSE)
}
\arguments{
\item{data}{A data frame with names matching those from PacFIN.}

\item{verbose}{A logical value specifying if you want messages printed to
the screen or not. The default is \code{FALSE}, which does not print them.}
}
\value{
A data frame with names similar to those exported by the survey-data
warehouse. No column are actually removed, columns are only added.
}
\description{
Change the column names of data extracted from PacFIN to match those
extracted from the survey-data warehouse.
}
