% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGrades.R
\name{checkGrade}
\alias{checkGrade}
\title{Check Sample Numbers for Grade}
\usage{
checkGrade(Pdata)
}
\arguments{
\item{Pdata}{A PacFIN \code{data.frame}}
}
\value{
todo: document what checkGrade returns.
}
\description{
Ensure that each sample number has only one grade of fish in it.
}
\author{
Kelli Faye Johnson
}
