% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGrades.R
\name{checkGrade}
\alias{checkGrade}
\title{Check Sample Numbers for Grade}
\usage{
checkGrade(Pdata)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}
}
\value{
todo: document what checkGrade returns.
}
\description{
Ensure that each sample number has only one grade of fish in it.
}
\author{
Kelli Faye Johnson
}
