% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanColumns.R
\name{cleanColumns}
\alias{cleanColumns}
\title{Change Column Names to Vdrfd}
\usage{
cleanColumns(data, use = c("vdrfd", "raw"))
}
\arguments{
\item{data}{A data frame with column names.}

\item{use}{Specify the column name category you would like to switch to.
\code{match.arg} is used to allow for partial matches.
Options include
`r paste0("``", eval(formals(cleanColumns)[["use"]]), "``", collapse = ", ")`.
\code{"vdrfd"} will return the column names used prior to 2020, the default.
\code{"raw"} will return the original column names as downloaded.}
}
\value{
An augmented data frame with new column names.
}
\description{
Change column names of comprehensive PacFIN columns to those
used in legacy code.
}
\author{
John R. Wallace, Kelli Faye Johnson
}
