% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPacFIN.R
\name{cleanPacFIN}
\alias{cleanPacFIN}
\title{Clean raw PacFIN data}
\usage{
cleanPacFIN(
  Pdata,
  keep_INPFC = lifecycle::deprecated(),
  keep_gears,
  keep_sample_type = c("", "M"),
  keep_sample_method = "R",
  keep_length_type,
  keep_age_method = NULL,
  keep_missing_lengths = lifecycle::deprecated(),
  keep_states = c("WA", "OR", "CA"),
  CLEAN = TRUE,
  spp = NULL,
  verbose = FALSE,
  savedir
)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{keep_INPFC}{\emph{Deprecated}. Areas are now defined using different methods.}

\item{keep_gears}{A character vector including only the gear types you want
to label as unique fleets. Order matters and will define fleet numbering.
If the argument is missing, which is the default, then all found gear groups
are maintained and ordered alphabetically. For more details see
\link{getGearGroup} that lists a web link for where you can find the
available gear groupings and how they link to \code{"GRID"} within your data.
GRID is a legacy term from PacFIN, now identified as
PACFIN_GEAR_CODE in the biological and fish ticket data, where
GR is short for gear and
ID is short for identification.
Typical entries will include character values such as \code{HKL}, \code{POT}, \code{TWL},
where the latter is short for all non-shrimp trawls and
\code{TWS} is shrimp trawls.
Other gear identification codes and their definitions include
\code{DRG} which is dredge gear,
\code{MSC} which is all other miscellaneous gear such as diving or river trawls,
\code{NET} which is all non-trawl net gear,
\code{NTW} which is non-trawl gear, and
\code{TLS} which is trolling gear.}

\item{keep_sample_type}{A vector of character values specifying the types of
samples you want to keep. The default is to keep \code{c("", "M")}. Available
types include market (M), research (R), special request (S), and
commercial on-board (C).}

\item{keep_sample_method}{A vector of character values specifying the types of
sampling methods you want to keep. The default is to keep \code{"R"}, which
refers to samples that were sampled randomly. Available types include
random (R), stratified (S), systematic (N), purposive (P), and special (X).
As of February 17, 2021,
Washington is the only state with a sample type of \code{""}, and it was limited
to two special samples of yelloweye rockfish.}

\item{keep_length_type}{A vector of character values specifying the types of
length samples to keep. There is no default value, though users will typically
want to keep \code{c("", "F", "A")}, but should also think about using
\code{c("", "F", "A", NA)}. Note that types other than those listed below
can be present, especially if you are dealing with a skate.
\code{A} is alternate length,
\code{D} is dorsal length,
\code{F} is fork length,
\code{S} is standard length, and
\code{T} is total length.}

\item{keep_age_method}{A vector of ageing methods to retain in the data. All fish
aged with methods other than those listed will no longer be considered aged.
A value of \code{NULL}, the default, will keep all ageing methods. However,
a vector of \verb{c("B", "BB", S", "", NA, 1, 2)} will keep all unaged fish and those
that were aged with break and burn and surface reads. You do not really need
to include such a verbose vector of values though because numbers are converted
to appropriate character codes in \link{getAge}. Therefore, something like
\code{c("B", "S")} would be sufficient to keep all break and burn and surface reads.}

\item{keep_missing_lengths}{\emph{Deprecated}. Just subset them using
\verb{is.na(Pdata[, 'length']) after running }cleanPacFIN` if you want to remove
lengths, though there is no need because the package accommodates keeping them in.}

\item{keep_states}{A vector of states that you want to keep, where each state
is defined using a two-letter abbreviation, e.g., \code{WA}. The default is to keep
data from all three states, \code{keep_states = c("WA", "OR", "CA")}.
Add \code{'UNK'} to the vector if you want to keep data not assigned to a state.}

\item{CLEAN}{A logical value used when you want to remove data from the input
data set. The default is \code{TRUE}. Where the opposite returns the original
data with additional columns and reports on what would have been removed.}

\item{spp}{A character string giving the species name to
ensure that the methods are species specific. Leave \code{NULL}
if generic methods work for your species.
Currently, sablefish is the only species with species-specific code.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors.
The printing doesn't affect any of the returned objects, instead,
printing makes it easier to see if the data has errors and what was
done to rectify them in the given function.
Sorry, but the default is to always print to the screen,
i.e., \code{verbose = TRUE}, so you do not miss out on anything.
This is standard practice for packages in the nwfsc-assess organization.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}
}
\value{
The input data filtered for desired areas and record types
specified, with added columns
\itemize{
\item year: initialized from SAMPLE_YEAR\cr
\item fleet: initialized to 1
\item fishery: initialized to 1
\item season: initialized to 1.  Change using \link{getSeason}
\item state: initialized from SOURCE_AGID.  Change using \link{getState}
\item length: length in mm, where \code{NA} indicates length is not available
\item lengthcm: floored cm from FORK_LENGTH when available, otherwise FISH_LENGTH
\item geargroup: the gear group associated with each GRID, from http://pacfin.psmfs.org/pacfin_pub/data_rpts_pub/code_lists/gr.
}
}
\description{
Clean raw PacFIN data to remove unsuitable samples if \code{CLEAN = TRUE} and
convert units of measured quantities to work with downstream functions.
Raw data are meant to be inclusive of everything from PacFIN so users can
explore all that is available, but this means that raw data will \strong{ALWAYS}
include information that is not appropriate for use in
US West Coast stock assessments.
}
\details{
The original fields in the returned data are left untouched,
with the exception of
\itemize{
\item \code{SEX}: modified using \link{getSex} and upon return will only include
character values such that fish with an unidentified sex are now \code{"U"}.
\item age: the best ages to use going forward rather than just the first age read.
}

\subsection{Workflow}{
If there are CalCOM samples to be integrated with PacFIN data, run \code{combineCalCOM}
first, otherwise run to \code{cleanPacFIN} as the first function in the workflow.
}

\subsection{Furthermore}{
The values created as new columns are for use by other functions in this package.
In particular, \code{fishyr} and \code{season} are useful if there are multiple
seasons (e.g., winter and summer, as in the petrale sole assessment), and the
year is adjusted so that "winter" occurs in one year, rather than across two.

The \code{fleet}, \code{fishery}, and \code{state} columns are meant for use in
stratifying the data according to the particulars of an assessment.
}
}
\examples{
data(XMPL.BDS)
XMPL.BDS[["FISH_WEIGHT_UNITS"]] <- "KG"
Pdata <- cleanPacFIN(XMPL.BDS, keep_length_type = unique(XMPL.BDS[, "FISH_LENGTH_TYPE"]))
NROW(XMPL.BDS) - NROW(Pdata)
# This time don't clean it
Pdata <- cleanPacFIN(XMPL.BDS, keep_length_type = unique(XMPL.BDS[, "FISH_LENGTH_TYPE"]),
  CLEAN = FALSE)
NROW(XMPL.BDS) - NROW(Pdata)
}
\seealso{
\link{getState}, \link{getSeason}
}
\author{
Andi Stephens
}
