% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineCalCOM.R
\name{combineCalCOM}
\alias{combineCalCOM}
\title{Combine CalCOM and PacFIN data}
\usage{
combineCalCOM(Pdata, CalCOM)
}
\arguments{
\item{Pdata}{Output from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}}.}

\item{CalCOM}{A data frame supplied, more than likely be E.J. Dick or
California Department of Fish and Wildlife containing biological samples
for flatfish species prior to 1990.}
}
\value{
Returns a combined dataset in PacFIN format.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Adds required PacFIN columns to CalCOM
data, initializing them to meaningful values as appropriate and translating
from the CalCOM values when they exist in a different format.
}
\details{
This function was made defunct in March 2021 in \href{https://github.com/pfmc-assessments/PacFIN.Utilities/commit/d5c0355795d10b945a9e28bfb05cb6811697c852}{commit d5c0355}
because CalCOM data were expected to be available in PacFIN and restored in
April 2023 after discovering that CalCOM data for flatfish prior to 1990 is
not currently available in PacFIN, as discussed in \href{https://github.com/pfmc-assessments/PacFIN.Utilities/issues/101}{issue #101} because the
early data are "in separate tables because they are bin samples (# of fish)
rather than cluster samples (weight based). PacFIN doesn't access those
tables."

Coding of data from CalCOM to PacFIN is done for dates, genders, and areas.
\code{PSMFC_AREA}s are derived from CalCOM \code{"PORT"}s.

Other data that are not reported in CalCOM datasets (e.g., \code{SAMPLE_TYPE} and
\code{SAMPLE_METHOD}) are initialized to typical default values kept when cleaning
data so that the CalCOM data aren't rejected by \code{\link[=cleanPacFIN]{cleanPacFIN()}}.
}
\seealso{
\itemize{
\item \code{\link[=cleanPacFIN]{cleanPacFIN()}}, which should be ran after this function
}
}
\author{
Melissa Haltuch, Andi Stephens, and Ian G. Taylor
}
