% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CombineCALCOM.R
\name{combineCALCOM}
\alias{combineCALCOM}
\title{Combine CALCOM and PacFIN data}
\usage{
combineCALCOM(Pdata, CALCOM)
}
\arguments{
\item{Pdata}{a PacFIN dataset}

\item{CALCOM}{a CALCOM dataset}
}
\value{
Returns a combined dataset in PacFIN format.
}
\description{
\code{combineCALCOM} adds required PacFIN columns to CALCOM data, initializing them
to meaningful values as appropriate, translating from the CALCOM values when
they exist in a different format.
}
\details{
Intended to be run before \code{\link{cleanPacFIN}}.

Date formats and genders are translated to PacFIN format, PSMFC_AREAs are derived
from CALCOM "PORT"s.

Other data that aren't reported in CALCOM datasets (e.g., SAMPLE_TYPE and
SAMPLE_METHOD) are initialized so that the CALCOM data aren't rejected by
\code{cleanPacFIN}.
}
\author{
Melissa Haltuch, Andi Stephens
}

