% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineCALCOM.R
\name{combineCalCOM}
\alias{combineCalCOM}
\title{Combine CalCOM and PacFIN data}
\usage{
combineCalCOM(Pdata, CalCOM)
}
\arguments{
\item{Pdata}{a PacFIN dataset}

\item{CalCOM}{a CalCOM dataset}
}
\value{
Returns a combined dataset in PacFIN format.
}
\description{
Adds required PacFIN columns to CalCOM data, initializing them
to meaningful values as appropriate, translating from the CalCOM values when
they exist in a different format.
}
\author{
Melissa Haltuch, Andi Stephens
}
