% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doDiags.R
\name{doRawDiags}
\alias{doRawDiags}
\title{Diagnostic plots and summaries for a raw PacFIN dataset pre-cleaning.}
\usage{
doRawDiags(Pdata, fname = NULL)
}
\arguments{
\item{Pdata}{a PacFIN dataset}

\item{fname}{a filename}
}
\description{
\code{doRawDiags} creates a set of diagnostic plots and summaries, writing pdfs and
a text file in addition to plotting onscreen and console.
}
\details{
Will create a filename from the species ID if one is not provided.
}

