% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSexRatio.R
\name{doSexRatio}
\alias{doSexRatio}
\title{Assign gender for unsexed fish in compiled Comp data.}
\usage{
doSexRatio(CompData, ratioU, maxsizeU, GTsizeU, savedir)
}
\arguments{
\item{CompData}{data that have already been aggregated by \code{getComps}.}

\item{ratioU}{the ratio to apply to fish less than maxsizeU}

\item{maxsizeU}{the size below which the sex ratio is assumed ratioU.}

\item{GTsizeU}{the size above which the ratio is assumed to be 1.0 (big mamas).}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}
}
\value{
Returns Comp data with unsexed fish now assigned sexes (the values for males and
females in the comps have increased), however the original columns
for unsexed fish remain unchanged.  The function that writes out the comps
\code{\link{writeComps}} sets these to zero.
}
\description{
doSexRatio determines gender for unsexed fish in Age or Length comps.

\subsection{Workflow}{
Sex ratios are assigned after the data is stratified by \code{\link{getComps}}
and before running \code{\link{writeComps}}.

\strong{If \code{\link{writeComps}} is run without running \code{doSexRatio}, all unsexed
fish will be discarded.}
}
}
\details{
Sex ratios may be assigned in one of four different ways.
}
\seealso{
\code{\link{getComps}}, \code{\link{writeComps}}
}
\author{
Andi Stephens, Kelli Johnson, and Chantel Wetzel
}
