% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EF1_Denominator.R
\name{EF1_Denominator}
\alias{EF1_Denominator}
\title{Calculate the denominator for the level-1 expansion factor.}
\usage{
EF1_Denominator(Pdata, Indiv_Wgts = TRUE, fa = 2e-06, fb = 3.5,
  ma = 2e-06, mb = 3.5, ua = 2e-06, ub = 3.5, verbose = TRUE,
  plot = FALSE)
}
\arguments{
\item{Pdata}{A PacFIN \code{data.frame}}

\item{Indiv_Wgts}{A logical value, controlling whether or not individual
fish weights should be generated using the specified weight-length relationship.}

\item{fa}{Female specific weight length coefficient.}

\item{fb}{Female specific weight length exponent.}

\item{ma}{Male specific weight length coefficient.}

\item{mb}{Male specific weight length exponent.}

\item{ua}{Unsexed weight length coefficient.}

\item{ub}{Unsexed weight length exponent.}

\item{verbose}{Report extra information.  Default:  TRUE}

\item{plot}{Create plots.  Default:  FALSE}
}
\value{
Additional columns are added to \code{Pdata}:
\itemize{
\item Wt_Sampled_1
\item Wt_Sampled_2
\item LW_Calc_Wt: individual weights predicted from the specified length-weight relationships.
\item Wt_Sampled_3
\item Wt_Sampled: the sample weight that will be used in subsequent analyses,
  where this is preferentially Wt_Sampled_1, with NAs replaced with values from Wt_Sampled_2,
  and NAs remaining replaced with values from Wt_Sampled_3.
\item Wt_Method: a \code{numeric} value denoting which method was used for \code{Wt_Sampled}.
}
}
\description{
\subsection{\code{\link{Workflow}}}{
\code{EF1_Denominator} is not run by the user.  It is a sub-function of 
\code{\link{getExpansion_1}}
}
}
\details{
The denominator of the level-1 expansion factor is the weight of the sampled
fish in a tow. The calcuation is done one in three ways, all of which
are returned, though the column used in subsequent analyses is based on
the hierarchical structure of:
\enumerate{
\item{\code{sum(Pdata$FEMALES_WGT + Pdata$MALES_WGT)} per unique \code{SAMPLE_NO}}
\item{\code{sum(Pdata$SPECIES_WGT)} across all clusters in a sample, where the
  species weight is determined from the \code{cluster_wt}}
\item{Calculate weights of males and females given the weight length relationship,
if the length of a fish does not exist, then the median weight of all
weighed fish in the sample is used.}
}
}
\seealso{
\code{\link{EF1_Numerator}}, \code{\link{getExpansion_1}}, \code{\link{getExpansion_2}}
}
\author{
Andi Stephens
}
