% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EF1_Numerator.R
\name{EF1_Numerator}
\alias{EF1_Numerator}
\title{Calculate the numerator for the first level expansion factor}
\usage{
EF1_Numerator(Pdata, verbose = TRUE, plot = FALSE)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fishieries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}

\item{plot}{Argument takes either a logical or character value specifying the file
name if you want to write the plots to a disk rather than printing them
to the console. If \code{plot = FALSE} no plots will be generated.
If printing to the disk the character value should end in \code{.png}
as \code{png()} is used to save the plotting device.}
}
\value{
A \code{Pdata} with additional columns, where
\code{Trip_Sampled_Lbs} is the sample weight in pounds.
}
\description{
Calculate the numerator for the first-level expansion factor, where
the numerator is the species-specific landing weight for a given sample.
Thus, if two clusters were sampled from a single trip,
they would both use the same landing weight.
}
\details{
Previously, \code{Trip_Sampled_Lbs} was calculated differently for each state.
For California, \code{Species_Percent_Sampled * TOTAL_WGT}.
For Oregon, \code{Pdata$EXP_WT} and if missing, the same as California.
For Washington, \code{Pdata$RWT_LBS}, \code{Pdata$TOTAL_WGT}, \code{RWT_LBS}, or
\code{median(Pdata$TOTAL_WGT)}.
Then, if all else failed, per-year, state-specific medians.

Now, PacFIN works hard behind the scenes to provide species-specific landing
weights for each sampled fish. Therefore, we no longer rely on code to
calculate a fabricated landing weight. Species-specific landing weights are
available in either \code{EXPANDED_SAMPLE_WEIGHT} or \code{WEIGHT_OF_LANDING_LBS}.
The former, is specific to Oregon and samples that do not provide an expanded
sample weight should not be used more than likely anyway.

\strong{todo}:
\itemize{
\item determine if we want to flag some bad samples in \link{cleanPacFIN}.
\item fix up the plotting and summary code
}
}
\seealso{
\link{getExpansion_1} calls this function.
}
\author{
Andi Stephens
}
