% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EF1_Numerator.R
\name{EF1_Numerator}
\alias{EF1_Numerator}
\title{Calculate the numerator for the first level expansion factor.}
\usage{
EF1_Numerator(Pdata, verbose = TRUE, plot = FALSE)
}
\arguments{
\item{Pdata}{A PacFIN \code{data.frame}}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors. The default value
is \code{TRUE}.}

\item{plot}{Argument takes either a logical or character value specifying the file
name if you want to write the plots to a disk rather than printing them
to the console. If \code{plot = FALSE} no plots will be generated.
If printing to the disk the character value should end in \code{.png}
as \code{png()} is used to save the plotting device.}
}
\value{
A \code{Pdata} with additional columns, where
  \code{Use_acs} is \code{all_cluster_sum} with NAs replaced by year-specific,
  state-specific medians;
  \code{Trip_Sampled_Lbs} is the sample weight in pounds;
}
\description{
\subsection{\code{\link{Workflow}}}{
\code{EF1_Numerator} is not run by the user.  It is a sub-function of 
\code{\link{getExpansion_1}}
}
}
\details{
\code{Trip_Sampled_Lbs} is calculated differently for each state:
\itemize{
\item {California}{ = \code{Species_Percent_Sampled} *
  \code{TOTAL_WGT};}
\item {Oregon}{ = \code{Pdata$EXP_WT}.  Where missing,
  use \code{Species_Percent_Sampled} * \code{TOTAL_WGT}, as for CA;}
\item {Washington}{ = \code{Pdata$RWT_LBS}, \code{Pdata$TOTAL_WGT},
  median(RWT_LBS), or median(\code{Pdata$TOTAL_WGT});}
\item {if all else fails}{ = use per-year, state-specific medians.}
}
}
\author{
Andi Stephens
}
