% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.matching.rows.R
\name{find.matching.rows}
\alias{find.matching.rows}
\title{Function find.matching.rows}
\usage{
find.matching.rows(file, table, findex = 1, tindex = 1, tcol = 2, round. = T)
}
\arguments{
\item{file}{Rows to be matched from \code{table}}

\item{table}{Superset of \code{file}}

\item{findex}{columns of \code{file} to use to match values in \code{tindex}}

\item{tindex}{columns of \code{table} to use to match values in \code{findex}}

\item{tcol}{Still mysterious.}

\item{round.}{if values are numeric, round.  Default:  TRUE.}
}
\description{
Takes two tables with a shared primary key, and
  returns the rows of the second table for which the
  keys match those of the first.
}
\details{
NOTE:  The way this is written assumes that the second table is a
        superset of the first (i.e., that each value is matched).
}
\seealso{
A sub-function of \code{\link{getExpansion_1}}
}
\author{
John R. Wallace (John.Wallace@noaa.gov), (revised) Andi Stephens, 2010.
}
