% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCatch.R
\name{formatCatch}
\alias{formatCatch}
\title{Format Catches from Long to Wide}
\usage{
formatCatch(catch, strat, valuename = "catch")
}
\arguments{
\item{catch}{A data frame with at least a column specifying
the year the catches took place (e.g., Year), a column for
each of the variables specified in strat, and a column that holds
the measured catches specified using the \code{valuename} argument.}

\item{strat}{A vector of colnames to stratify the catch over}

\item{valuename}{The column name that contains the 
catch values in \code{catch}}
}
\description{
Transform a long data frame of catches to a wide data frame.
The colnames of the wide data frame will be in the format
needed for the statification of stage-2 expansions of
composition data, i.e., \code{getExpansion_2}.
}
