% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCatch.R
\name{formatCatch}
\alias{formatCatch}
\title{Format catches from long to wide}
\usage{
formatCatch(
  catch,
  strat,
  yearname = "^Year|^Yr|Landing_Y|Sample_Y",
  valuename = "ROUND_WEIGHT_LBS"
)
}
\arguments{
\item{catch}{A data frame with at least a column specifying
the year the catches took place
(e.g., LANDING_YEAR is the column name upon download from PacFIN),
a column for variable(s) specified in strat, and
a column that holds the measured catches named
(e.g., ROUND_WEIGHT_LBS is the column name upon download from PacFIN).}

\item{strat}{A vector of column names to stratify the output over.
For example, if you wish to summarize by ageing method, then the argument
would look like \code{strat = "agemethod"}.
Or if you want to look at fleets and gear, then use
\code{strat = c("fleet","usegear")}.
todo: deprecate defaults}

\item{yearname}{A character string used to search for year in \code{catch}.
Ensure that the string is unique enough that it will only return one value.
Though, if you would like catch by year and month you can pass a single
character value that includes both of these, for example \verb{_YEAR|_MONTH}.
This is underhanded and not what the original functionality of this function
intended, but feel free to use it.
This is not case specific.}

\item{valuename}{The column name that contains the
catch values in the data frame \code{ROUND_WEIGHT_LBS} is the default and the
name of the column that contains landed weight converted to round weight
in lbs. The conversion factor from landed weight to round weight is
weight_of_catch * NVL(conversion_factor,1) and 1 mt equals 2,204.62 lbs.
This is not case specific.}
}
\value{
A data frame in wide format with \code{yearname} being the first column
ordered from oldest to newest. Stratification columns follow and the names
are separated with a full stop when more than one stratification is provided.
Values in the rows will be the \code{valuename} summed across all available values.
If there are no relevant values for a given stratification-year combination,
then the entry will be zero.
}
\description{
Transform a long data frame of catches to a wide data frame
using \link[tidyr:pivot_wider]{tidyr::pivot_wider}.
The column names of the wide data frame will be in the format
needed for the stratification of stage-2 expansions of
composition data, i.e., \link{getExpansion_2}.
}
\author{
Kelli F. Johnson
}
