% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAge.R
\name{getAge}
\alias{getAge}
\title{Get age information for each fish}
\usage{
getAge(Pdata, verbose = FALSE, keep, col.bestage = "FISH_AGE_YEARS_FINAL")
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors.
The printing doesn't affect any of the returned objects, instead,
printing makes it easier to see if the data has errors and what was
done to rectify them in the given function.
Sorry, but the default is to always print to the screen,
i.e., \code{verbose = TRUE}, so you do not miss out on anything.
This is standard practice for packages in the nwfsc-assess organization.}

\item{keep}{A vector of values that represent what you want to keep.
Values of \code{NA}, \code{""}, and numeric values are acceptable.
Often, it is helpful to run `unique(, useNA = "always") on the relevant
data prior to running the function to see what types are present.}

\item{col.bestage}{A character value noting the column name that contains the best age
to use going forward for a given fish. Over time, this has been
\code{FISH_AGE_YEARS_FINAL} and \code{FINAL_FISH_AGE_IN_YEARS} for PacFIN.}
}
\value{
A vector of ages in years with a length the same as the number
of rows present in the input data frame \code{Pdata}.
If all ages are \code{NA}, then a vector of \code{NA} values is quickly returned.
}
\description{
Get age information for each fish, or row of the data.
This functionality was originally in cleanAges, moved to
cleanPacFIN, and is now compartmentalized here, but called in
\link{cleanPacFIN} and stored in the \code{Age} column.
}
\details{
Age types listed in keep (or derivatives of those) are used to filter
for ages that should be kept, i.e., not returned as \code{NA}.
If \code{FISH_AGE_YEARS_FINAL} doesn't have an entry, where this should be the
best age available, then the age reads in \code{age1}, \code{age2}, ... are
averaged over and rounded to the next integer
(i.e., \link{ceiling}) if their methods are acceptable.
When \code{FISH_AGE_YEARS_FINAL} does have an entry, this age is kept if one of
the methods in \code{AGE_METHOD} is listed in \code{keep}. It is currently unclear
how \code{FISH_AGE_YEARS_FINAL} is determined because it doesn't always equal
the value from the first age reader. Sometimes, it is \code{ceiling(mean())},
other times it appears to default to break and burn, and other random combos.
We use the best judgment of those who created the column that they know what
they are doing, but please feel free to report errors in PacFIN to the
github issues or the package maintainer after contacting the state representative.

The input argument of \code{keep} can contain a mix of numeric and character values,
but all values will be converted to character values. See below for conversions,
where it is easiest to just supply a vector of values shown on the left before
the double equal sign because all values in your data set to the right will be
converted prior to searching for those in \code{keep}.
For example, \code{keep = c("B", "L")} will keep all \code{"BB"}, \code{1}, and \code{6} entries as well.
\itemize{
\item "B" == 1, "BB", or break and burn
\item "S" == 2, "SR" (California), or surface read
\item 3 == scales (Oregon data)
\item "T" == 4, "TS" (California), "X" (Washington), or thin section
\item "O" == 5 or optical scanner
\item "L" == 6 or age derived from length
\item 9 == unable to read or age
\item U == unknown
}

The final age assigned to a fish when there are multiple age reads depends on the
ageing lab. See issue #49 for a semi-detailed discussion. Sometimes, reconciling
happens where the best age will not be equal to any age read from a specific reader
but the readers get together to determine the best read.
Washington uses the following order of operations to assign the best age given
there are multiple reads. Note, that they will never report more than three reads
because that is the limitation of their database.
\itemize{
\item age1 if only one read;
\item randomly chosen age between age-read 1 and age-read 2 if the difference between
the two reads is less than three;
\item if the difference between two reads is more than three, then the best age is
manually keyed in from that designated by a human;
\item if three reads are equal, then the value is automatically assigned; and
\item if there are three reads and they are not equal then an age reader must choose
the final value.
}
}
