% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArea.R
\name{getArea}
\alias{getArea}
\title{Get the general area that the information was taken from}
\usage{
getArea(Pdata, verbose = FALSE)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors.
The printing doesn't affect any of the returned objects, instead,
printing makes it easier to see if the data has errors and what was
done to rectify them in the given function.
Sorry, but the default is to always print to the screen,
i.e., \code{verbose = TRUE}, so you do not miss out on anything.
This is standard practice for packages in the nwfsc-assess organization.}
}
\value{
A vector of character values, where NA entries are fine with respect to
what Kelli F. Johnson determined as being 'fine' and character entries attempt to
specify the poor area from which the record originated from. For example,
"Puget Sound" clearly indicates that the record was from the Puget Sound, an area
that most stocks do not include in their spatial footprint.
This vector will be of the same length as the number of rows of the input data frame.
}
\description{
Get the general area that the information (i.e., row of data) was taken from.
Thankfully, PacFIN contains records from all areas, but unfortunately,
we do not always want to use data from all of these areas in the stock assessment.
Thus, this function helps to identify which data should be included based on
standard filtering. More in-depth filtering will be case specific and should
be performed by the user.
}
\details{
Washington Department of Fish and Wildlife (WDFW) uses management regions, also
known as Marine Fish-Shellfish Management and Catch Reporting Areas, to define
the area in which landings occurred and sampled fish come from.
These management regions are numeric and fall within the range of \code{1:99}.
Not all numeric values within this range are actually used, but they are helpful
in that WDFW uses their management regions to define the PSMFC area in which
landings occurred because PSMFC area is not a "recorded" value on fish tickets.
See Table 3 on page 15 of
\href{https://wdfw.wa.gov/sites/default/files/publications/01754/wdfw01754.pdf}{Tsou et al., 2015}
for a table that converts WDFW management area to PSMFC area.
I had a difficult time finding a map of the 99 management areas, though a partial
map of the Puget Sound areas is provided in Figure 3 on page 26 of the
Puget Sound Groundfish Management Plan by
\href{https://wdfw.wa.gov/sites/default/files/publications/00927/wdfw00927.pdf}{Palsson et al., 1998}
and an arcgis layer of the areas within the Strait of Juan de Fuca and the Puget Sound
was posted by \href{https://www.arcgis.com/apps/webappviewer/index.html?id=13a09a6daae7408383fe3617ced75259}{SITC}.
Both of these resources were helpful in understanding how WDFW management areas 20-29
correspond to PSMFC area 4a and should be attributed to the Puget Sound.
Where, 4a corresponds to everything east of Cape Flattery
(i.e., the most easterly point of Washington state).

Foreign samples, those coming from outside of the EEZ, could be in your data.
The PacFIN documentation for
\href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/ar_tree.txt}{areas}
identifies areas ZY or ZZ or 9\link{GMYZ} as those outside of the EEZ.
Additionally, 3D and 3N make up VC, which is Canadian catches within the Vancouver area.
}
\seealso{
\link{cleanPacFIN} uses this function to get the general area and alert users
which rows should be excluded from their data.
}
