% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComps.R
\name{getComps}
\alias{getComps}
\title{Aggregate length, age, or age-at-length composition data by strata}
\usage{
getComps(
  Pdata,
  strat = NULL,
  Comps = c("LEN", "AGE", "AAL"),
  defaults = c("fleet", "fishyr", "season"),
  towid = c("SAMPLE_NO"),
  weightid = "Final_Sample_Size_L"
)
}
\arguments{
\item{Pdata}{A data frame returned from \code{\link[=PullBDS.PacFIN]{PullBDS.PacFIN()}} containing
biological samples. These data are stored in the Pacific Fishieries
Information Network (PacFIN) data warehouse, which originated in 2014 and
are pulled using sql calls.}

\item{strat}{A vector of column names to stratify the output over. For
example, if you wish to summarize by ageing method, then the argument
would look like \code{strat = "agemethod"} or, if you want to look at fleets
and gear, \code{strat = c("fleet", "usegear")}.}

\item{Comps}{The type of composition data to create. See the function call
for the available options. The first option listed is the default, which
creates conditional age-at-length compositions by adding both \code{lengthcm}
and \code{Age} to the grouping structure.}

\item{defaults}{The default stratification columns which will typically be
left at their default value of \code{c("fleet", "fishyr", "season")}.}

\item{towid}{A vector of character values providing the column names that
generate a unique id for each sample. The default is \code{"SAMPLE_NO"} but you
can pass a vector where all of the columns identified will be united to
create a unique identifier for the tow/sample.}

\item{weightid}{A character value giving the column name that holds the value
to be summed for each type and strata, such as \code{weightid = "Final_Sample_Size_L"}, if you are interested in length data. Instead of
the default, you might want to weight each stratification by the final
sample sizes found for age data, i.e., \code{weightid = "Final_Sample_Size_A"}.
You can pass \code{weightid} any column that exists in \code{Pdata}. Legacy code did
not allow you to name the column and \code{Final_Sample_Size} was hard coded in
the function.}
}
\value{
A long data frame of weights for each grouping structure. Columns identifying
the grouping structures will come first, followed by columns with sample
sizes per group by sex category. The documentation for these sample size
columns is sparse because this function is set to be deprecated next cycle
and replaced with a simplified path to writing composition data. But,
information is present for males (abbreviated with an M), females
(abbreviated with an F), unsexed (abbreviated with a U). If no sex is
provided then it is assumed all are unsexed and will be returned as such.
}
\description{
\code{getComps()} first sets up the local environment and returns summaries by
groupings.
}
\details{
This function uses the expansions created in \code{\link[=getExpansion_1]{getExpansion_1()}} and
\code{\link[=getExpansion_2]{getExpansion_2()}} using \code{"weightid"}. Thus, whatever column name is passed
to \code{weightid} will be used as the weight for each group. See the
documentation for the pass-through arguments, i.e., \code{...}
}
\seealso{
\itemize{
\item \code{\link[=getExpansion_2]{getExpansion_2()}} should be run before this function.
\item \code{\link[=writeComps]{writeComps()}} can be run on the output from this function.
}
}
\author{
Andi Stephens, Kelli F. Johnson
}
