% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComps.R
\name{getComps}
\alias{getComps}
\title{Aggregate composition data by length, age, or age-at-length according to the
given stratification.}
\usage{
getComps(
  Pdata,
  strat = NULL,
  Comps = "AAL",
  defaults = c("fleet", "fishyr", "season"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{strat}{A vector of column names to stratify the output over.
For example, if you wish to summarize by ageing method, then the argument
would look like \code{strat = "agemethod"}.
Or if you want to look at fleets and gear, then use
\code{strat = c("fleet","usegear")}.
# todo: deprecate defaults}

\item{Comps}{The type of composition data to create. Options are length
(\code{'LEN'}, age (\code{'AGE'}), or conditional age-at-length (\code{'AAL'}).}

\item{defaults}{The default stratification columns
which will typically be left at their default value of
\code{c('fleet', 'fishyr', 'season')}.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors. The default value
is \code{TRUE}.}

\item{...}{Pass additional arguments to \code{getcomps_long}, such as
\code{dropmissing = FALSE} where the default behaviour is 
\code{dropmissing = TRUE}}
}
\value{
A dataframe with composition data specific to the type specified
  in \code{Comps} for males, females, and unsexed records.
}
\description{
\subsection{\code{\link{Workflow}}}{
\code{getComps} is run subsequently to \code{\link{getExpansion_2}}.
}
}
\details{
The aggregation is of the \code{Pdata$Final_Sample_Size} column value,
  which should be set to the desired expansion:
  
  \code{Pdata$Final_Sample_Size = Pdata$Expansion_Factor_1} 
  
  or
  
  \code{Pdata$Final_Sample_Size = Pdata$Expansion_Factor_1 * Pdata$Expansion_Factor_2}
  
  The default stratification is by fleet, fishyr, and season.
  Columns \code{lengthcm}, \code{age} or both are added
  depending on the \code{Comps} argument.
}
\author{
Andi Stephens, Kelli Faye Johnson
}
