% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDB.R
\name{getDB}
\alias{getDB}
\title{Get a connection to a database and return a query}
\usage{
getDB(
  sql,
  datasourcename = "PacFIN",
  username = getUserName(datasourcename),
  password,
  asis = FALSE
)
}
\arguments{
\item{sql}{A valid SQL statement.}

\item{datasourcename}{The sql data base name that is stored on your
computer. The default is \code{"PacFIN"} which will pull data from
the Pacific Fisheries Information Network.}

\item{username}{The username you would like to use to access the database.
The default is to generate \code{username} from your computer name where
there are rules for each supported database. Sometimes this fails because
of legacy rules that are unknown to the development team. Please email the
maintainer of this package if you need more behind-the-scenes
functionality here.}

\item{password}{Your password for the database of interest. Use
\code{\link[=ask_password]{ask_password()}} to be prompted for your password. Please do not share this
password with anyone or push code to a repository that has your password
saved in it.}

\item{asis}{A logical, specifying whether or not to convert columns,
as in \code{\link[utils]{read.table}}.}
}
\value{
A data frame resulting from querying the database or an
error message because the main call is wrapped in
\code{\link[base]{tryCatch}}.
}
\description{
Get a connection established to a database and return the information desired
in the specific query.
}
\author{
John R. Wallace, Kelli F. Johnson
}
