% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpansion_1.R
\name{getExpansion_1}
\alias{getExpansion_1}
\title{First level expansion for composition data to accout for unsampled
fish in a tow.}
\usage{
getExpansion_1(
  Pdata,
  maxExp = 0.95,
  Indiv_Wgts = TRUE,
  Exp_WA = TRUE,
  fa = NA,
  fb = NA,
  ma = NA,
  mb = NA,
  ua = NA,
  ub = NA,
  verbose = TRUE,
  plot = FALSE
)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{maxExp}{The maximum expansion factor (either a number or a quantile).}

\item{Indiv_Wgts}{A logical value, controlling whether or not individual
fish weights should be generated using the specified weight-length relationship.}

\item{Exp_WA}{Default FALSE.  If TRUE, expand the WA samples.}

\item{fa}{Female specific weight length coefficient.}

\item{fb}{Female specific weight length exponent.}

\item{ma}{Male specific weight length coefficient.}

\item{mb}{Male specific weight length exponent.}

\item{ua}{Unsexed weight length coefficient.}

\item{ub}{Unsexed weight length exponent.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors. The default value
is \code{TRUE}.}

\item{plot}{Argument takes either a logical or character value specifying the file
name if you want to write the plots to a disk rather than printing them
to the console. If \code{plot = FALSE} no plots will be generated.
If printing to the disk the character value should end in \code{.png}
as \code{png()} is used to save the plotting device.}
}
\value{
A \code{data.frame} where all of the original columns in
  \code{Pdata} remain unaltered but additional columns are added.
}
\description{
\subsection{\code{\link{Workflow}}}{
\code{getExpansion_1} is run after \code{\link{cleanPacFIN}}.
}
}
\details{
\code{getExpansion_1} calls \code{\link{EF1_Numerator}} and
  \code{\link{EF1_Denominator}} (i.e., the weight of sampled fish and
  the weight of all fish of the respective species in the tow) and returns
  their ratio.
}
\section{Expansion caveats}{

There is one manual step in the workflow. 
After running the expansion functions, data columns Expansion_Factor_1 and 
Expansion_Factor_2 are available to use in manually setting the Final_Expansion_Factor. 
\itemize{
\item{Age data are expanded separately from lengths.}
\item{WA fish are generally only expanded using Expansion_Factor_2.}
\item{Other expansions are the product of Expansion_Factor_1 * Expansion_Factor_2}
\item{For age-at-length comps, set Final_Expansion_Factor to 1.  Each fish represents only itself.}
}
}

