% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpansion_1.R
\name{getExpansion_1}
\alias{getExpansion_1}
\title{First-Stage Expansion for Composition Data}
\usage{
getExpansion_1(
  Pdata,
  maxExp = 0.95,
  Indiv_Wgts = TRUE,
  Exp_WA = TRUE,
  fa = NA,
  fb = NA,
  ma = NA,
  mb = NA,
  ua = NA,
  ub = NA,
  verbose = FALSE,
  plot = FALSE
)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{maxExp}{The maximum expansion factor
(either a number or a quantile) for building expansions.
Typically, the default is 0.95.
Set \code{maxExp = Inf} to see largest values.}

\item{Indiv_Wgts}{A logical value, controlling whether or not individual
fish weights should be generated using the specified weight-length relationship.}

\item{Exp_WA}{Default FALSE.  If TRUE, expand the WA samples.}

\item{fa}{Female specific weight length coefficient.}

\item{fb}{Female specific weight length exponent.}

\item{ma}{Male specific weight length coefficient.}

\item{mb}{Male specific weight length exponent.}

\item{ua}{Unsexed weight length coefficient.}

\item{ub}{Unsexed weight length exponent.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors. The default value
is \code{TRUE}.}

\item{plot}{Argument takes either a logical or character value specifying the file
name if you want to write the plots to a disk rather than printing them
to the console. If \code{plot = FALSE} no plots will be generated.
If printing to the disk the character value should end in \code{.png}
as \code{png()} is used to save the plotting device.}
}
\value{
A \code{data.frame} where all of the original columns in
\code{Pdata} remain unaltered but additional columns are added.
In particular columns starting with
Expansion_Factor_1 are available for setting the Final_Expansion_Factor.
}
\description{
First-stage expansions account for unsampled fish in the smallest
measured unit. Where, in PacFIN data, the smallest unit is that
which is available to the port or dockside sampler, so more than
likely a trip rather than a tow as would be with survey data.
}
\details{
The workflow is to run \code{getExpansion_1} after \code{\link{cleanPacFIN}},
which assures that all of the necessary columns are available and that the
data are in the correct units.
\code{getExpansion_1} calls
\code{\link{EF1_Numerator}} and \code{\link{EF1_Denominator}}
(i.e., the weight of sampled fish and
the weight of all fish of the respective species in the tow)
and returns their ratio.
}
\section{Expansion}{

\itemize{
\item{Age data are expanded separately from lengths.}
\item{WA fish are generally only expanded using Expansion_Factor_2.}
\item{Other expansions are the product of Expansion_Factor_1 * Expansion_Factor_2}
\item{For age-at-length comps, set Final_Expansion_Factor to 1 because
  each fish represents only itself.}
}
}

\seealso{
Called within this function are
\code{\link{EF1_Numerator}} \code{\link{EF1_Denominator}}.
Called after this function should be \code{\link{getExpansion_2}}.
}
