% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGearGroup.R
\name{getGearGroup}
\alias{getGearGroup}
\title{Create column for gear called \code{geargroup} according to PacFIN gears}
\usage{
getGearGroup(Pdata, spp = NULL, verbose = TRUE)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{spp}{A character string giving the species name to
ensure that the methods are species specific. Leave \code{NULL}
if generic methods work for your species.
Currently, sablefish is the only species with species-specific code.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors.
The printing doesn't affect any of the returned objects, instead,
printing makes it easier to see if the data has errors and what was
done to rectify them in the given function.
Sorry, but the default is to always print to the screen,
i.e., \code{verbose = TRUE}, so you do not miss out on anything.
This is standard practice for packages in the nwfsc-assess organization.}
}
\value{
A modified data frame where an additional column labeled
\code{geargroup} is added to \code{Pdata}. No original columns are modified
in the process.
}
\description{
Data from the PacFIN
\href{http://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/gr.txt}{gear table}
is used to create a column in \code{Pdata} called \code{geargroup}, where
\code{Pdata$GRID} is recoded to \code{geargroup} according to the gear table.
}
\examples{
ex <- getGearGroup(data.frame(GRID = c("PRT", "BMT", "FPT")), verbose = FALSE)
table(ex)
testthat::expect_equal(ex[ex[, "geargroup"] == "POT", "GRID"], "FPT")

}
\seealso{
Reverse dependency of \link{cleanPacFIN}
}
\author{
Andi Stephens
}
