% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLength.R
\name{getLength}
\alias{getLength}
\title{Get length for PacFIN commercial samples}
\usage{
getLength(Pdata, verbose = TRUE, keep)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors.
The printing doesn't affect any of the returned objects, instead,
printing makes it easier to see if the data has errors and what was
done to rectify them in the given function.
Sorry, but the default is to always print to the screen,
i.e., \code{verbose = TRUE}, so you do not miss out on anything.
This is standard practice for packages in the pfmc-assessments organization.}

\item{keep}{A vector of values that represent what you want to keep.
Values of \code{NA}, \code{""}, and numeric values are acceptable.
Often, it is helpful to run `unique(, useNA = "always") on the relevant
data prior to running the function to see what types are present.}
}
\value{
A vector of lengths in millimeters. Values of \code{NA} indicate that
the length should not be used in down-stream calculations.
}
\description{
Get length for PacFIN commercial samples in millimeters.
Lengths are converted for skates from type R and A to actual lengths.
Other, more standard lengths, are filtered for types in \code{keep}.
}
