% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getState.R
\name{getState}
\alias{getState}
\title{Create a state column based on input column specified in \code{source}}
\usage{
getState(
  Pdata,
  source = c("SOURCE_AGID", "PSMFC_ARID", "SAMPLE_AGENCY", "AGID", "AGENCY_CODE"),
  verbose = FALSE
)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{source}{The column name where state information is located in
\code{Pdata}. See the function call for options, where only the first
value will be used.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors.
The printing doesn't affect any of the returned objects, instead,
printing makes it easier to see if the data has errors and what was
done to rectify them in the given function.
Sorry, but the default is to always print to the screen,
i.e., \code{verbose = TRUE}, so you do not miss out on anything.
This is standard practice for packages in the nwfsc-assess organization.}
}
\description{
Create a categorical column that specifies which state each row is from.
}
\examples{
data(XMPL.BDS)
invisible(getState(XMPL.BDS, verbose = FALSE))
testthat::expect_equivalent(
  table(getState(XMPL.BDS, verbose = FALSE)[, "state"])[3],
  17292)

}
\seealso{
\link{cleanPacFIN} calls this function
}
