% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getState.R
\name{getState}
\alias{getState}
\title{Create a state field from \code{source} for the \code{data.frame}}
\usage{
getState(Pdata, source = "SOURCE_AGID", CLEAN = TRUE, keepPW = FALSE,
  verbose = TRUE)
}
\arguments{
\item{Pdata}{A PacFIN \code{data.frame}}

\item{source}{The column name where state information is located in
\code{Pdata}. Typical options are \code{'SOURCE_AGID'},
\code{'PSMFC_ARID'}, or \code{'SAMPLE_AGENCY'}.}

\item{CLEAN}{A logical value indicating if records for which state
cannot be assigned are removed or if just summary stats on what would
have been removed are generated.}

\item{keepPW}{A logival value indicating if records for Pacific Whiting
should be removed or kept in as their own state.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors. The default value
is \code{TRUE}.}
}
\description{
\code{Pdata}, where \code{Pdata} is a \code{data.frame} from
PacFIN.#' \subsection{\code{\link{Workflow}}}{
\code{getState} is run by \code{\link{cleanPacFIN}}.
}
}

