% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWLpars.R
\name{getWLpars}
\alias{getWLpars}
\title{Calculate Weight-Length Relationship}
\usage{
getWLpars(data, verbose = FALSE)
}
\arguments{
\item{data}{A data frame containing empircal weights and lengths
from sampled fish.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors. The default value
is \code{TRUE}.}
}
\value{
A data frame of weight-length parameters by sex.
Parameters A and B are in the appropriate units to input
into Stock Synthesis Wtlen_1_Fem and Wtlen_2_Fem, or
Wtlen_1_Mal and Wtlen_1_Mal, parameters in the 
control file.
}
\description{
Estimate parameters of the weight-length relationship for each
sex and all sexes combined, where the latter includes unsexed fish.
}
\details{
todo: Return model diagnostics if verbose. This will entail
saving each model in an object and then extracting the pars
via getline rather than doing it all in one step.
}
\author{
Kelli Faye Johnson
}
