% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PacFIN.Utilities-deprecated.R
\name{PacFIN.Utilities-deprecated}
\alias{PacFIN.Utilities-deprecated}
\alias{checkINPFC}
\alias{ls_INPFC}
\title{Deprecated function in package \pkg{PacFIN.Utilities}}
\usage{
checkINPFC(...)

ls_INPFC(...)
}
\description{
The functions listed below are deprecated and
will be removed from the package in the near future.
When possible, alternative functions with similar functionality are
available at \code{help("PacFIN.Utilities-deprecated")}.
}
\section{checkINPFC}{

Functions related to INPFC areas have been deprecated in favor of PSMFC areas.
}

\section{ls_INPFC}{

Functions related to INPFC areas have been deprecated in favor of PSMFC areas.
}

\keyword{internal}
