% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.col.R
\name{paste.col}
\alias{paste.col}
\title{Utility function.}
\usage{
paste.col(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
Pasted columns
}
\description{
\subsection{\code{\link{Workflow}}}{
Used by \code{\link{find.matching.rows}}, not intended to be used otherwise.
}
}
\details{
{
Converts each row in the input dataframe to a string, 
"pasting" the columns together.
}
}
