% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCleaned.R
\name{plotCleaned}
\alias{plotCleaned}
\title{Plot cleaned fishery age or length composition data.}
\usage{
plotCleaned(Pdata, length = TRUE)
}
\arguments{
\item{Pdata}{A matrix of age or length data returned from
\code{\link{cleanAges}} or \code{\link{cleanPacFIN}}.}

\item{length}{A logical value defining whether or not \code{Pdata} is length-
or age-composition data.}
}
\value{
A plot of either 3 x 2 or 2 x 2 depending on whether or not unique
  gear groups exist. Left column are plots of sample size by type and right
  column is boxplots of either length or age and depth by year.
}
\description{
Plot cleaned fishery age or length composition data.
}
\author{
Andi Stephens, Kelli Faye Johnson
}
\seealso{
\code{\link{cleanAges}}, \code{\link{cleanPacFIN}}
}

