% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRawData.R
\name{plotRawData}
\alias{plotRawData}
\title{Diagnostic plots and summaries for a raw PacFIN dataset pre-filtering.}
\usage{
plotRawData(rawData, fname = NULL)
}
\arguments{
\item{rawData}{an unfiltered PacFIN dataset}

\item{fname}{A filename with the appropriate extension, used to save the
function output to the disk. For example, \code{LenComps.csv} or
\code{file.path(getwd(), "SampleSize.csv")}. Full, relative, or simple paths are
allowed because the argument is used as is, i.e., not redirected to a
directory different than \code{\link[=getwd]{getwd()}}. Note that various functions within
pacfintools have different default values for this input argument. If
\code{NULL} in \code{writeComps}, then the resulting file name will be based on what
type of composition data is being generated, i.e., \code{PacFIN_lengths.out},
or \code{PacFIN_ages.out} for length or age data, respectively.}
}
\description{
\code{plotRawData} creates a set of diagnostic plots and summaries, writing
pdfs and a text file in addition to plotting onscreen and console.
}
\details{
\subsection{Workflow}{
Run \code{plotRawData} to visualize and summarize the PacFIN data prior to
running \code{\link{cleanPacFIN}}.
}

Will create a filename from the species ID if one is not provided.
}
