% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStrat.R
\name{plotStrat}
\alias{plotStrat}
\title{Plot Length Distributions with Aged and All Fish}
\usage{
plotStrat(data, dir, ylimperc = 0.65, npages = 11, dolbins)
}
\arguments{
\item{data}{A data frame}

\item{dir}{The directory you want to print the plots}

\item{ylimperc}{The percent of the maximum value in the data 
set that you want displayed on the y axis.}

\item{npages}{The number of pages you want the plots split over.}

\item{dolbins}{The length bins}
}
\description{
todo: document
}
