% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullBDS.PacFIN.R
\name{PullBDS.PacFIN}
\alias{PullBDS.PacFIN}
\title{Pull biological data from the PacFIN database}
\usage{
PullBDS.PacFIN(
  pacfin_species_code,
  username = getUserName("PacFIN"),
  password,
  savedir = getwd(),
  verbose = TRUE
)
}
\arguments{
\item{pacfin_species_code}{A character string for the PacFIN
species code that you desire to pull data for. Known also as
\code{"SPID"} in the PacFIN database if using legacy sql scripts.
An example for sablefish would be
\code{pacfin_species_code = "SABL"}.}

\item{username}{The username you would like to use to access data from
\code{datasourcename}. The default is to generate \code{username}
from your computer name.}

\item{password}{Your password for the \code{datasourcename}.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors.
The printing doesn't affect any of the returned objects, instead,
printing makes it easier to see if the data has errors and what was
done to rectify them in the given function.
Sorry, but the default is to always print to the screen,
i.e., \code{verbose = TRUE}, so you do not miss out on anything.
This is standard practice for packages in the nwfsc-assess organization.}
}
\value{
An RData object is saved to the disk and the pulled data
are returned as a data frame.
The saved data can be read back in using \link{load}, but note that
the object is named inside of the \code{.RData} file, and thus,
the object will retain the name of \code{bds.pacfin} within your work space.
In February of 2021, the returned object was renamed from \code{out}
to \code{bds.pacfin} in an attempt to be more clear.

Be aware that the data are changed from a long table to a wide table
using the combination of unique \code{FISH_IDs} in a given year and \code{AGE_SEQUENCE_NUMBER}.
Users are warned if there are non-unique \code{FISH_IDs} in a given year as these data
are automatically removed from the data frame. So, it is good to have \code{verbose = TRUE}.
This change from long to wide allows for rows equating to a single fish with columns
containing information about all measurements for that fish. Multiple age reads and
information about those reads such as age reader will be in the columns.
So, if your data had at most three duplicate reads, then you will see numbers up
to \code{3} pasted onto the end of relevant column names. This is to help you match duplicate
reads to the relevant information. All information from age reader #1 will be in
a column with no tag, i.e., \code{age} instead of \code{age2}.

\code{AGE_COUNT} is a somewhat cryptic column name and does not always make sense
when compared to \code{AGE_SEQUENCE_NUMBER}. It was determined that the former is
useful to identify how many potential agers were exposed to this fish.
For example, if \code{AGE_SEQUENCE_NUMBER} has a maximum value of three for a
given \code{FISH_ID}, then you can expect \code{AGE_COUNT} to be three for all three
rows in the PACFIN database for that fish. This is not always true though.
Sometimes, not all \code{AGE_SEQUENCE_NUMBER}s are present and they can skip
numbers for a given \code{FISH_ID}, and in this case, \code{AGE_COUNT} will be the
maximum \code{AGE_SEQUENCE_NUMBER} for a given \code{FISH_ID}.

Some column names have been renamed, these pertain to the reading of ages.
For example \code{AGE_IN_YEARS} is now \code{age}, which can be followed by a number if the age
was a second, third, fourth, ... read. Numeric values signify columns like \code{age}
where the information was gleaned from another row with the same unique FISH_ID
in a given year. Having this information of double reads by FISH_ID will facilitate
the estimation of ageing error.
}
\description{
Pull biological data from the
Comprehensive Biological Data (comprehensive_bds_comm) table
in the PacFIN database (PACFIN_MARTS).
Data are in their raw form and should be cleaned either by hand or
using \link{cleanPacFIN}.
}
\examples{
\dontrun{
# Enter password in place of xxx below
mypw <- "xxx"
pd <- PullBDS.PacFIN(pacfin_species_code = "POP", password = mypw)
}

}
\seealso{
\itemize{
\item \link{cleanColumns} to change to legacy column names
\item \link{cleanPacFIN} to manipulate and subset the returned object
}
}
\author{
John R. Wallace, Kelli Faye Johnson
}
