% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullBDS.PacFIN.R
\name{PullBDS.PacFIN}
\alias{PullBDS.PacFIN}
\title{Pull Catch Data from PacFIN Database}
\usage{
PullBDS.PacFIN(
  pacfin_species_code,
  username = getUserName("PacFIN"),
  password,
  savedir = getwd()
)
}
\arguments{
\item{pacfin_species_code}{A character string for the PacFIN
species code that you desire to pull data for. Known also as
\code{"SPID"} in the PacFIN database if using legacy sql scripts.
An example for sablefish would be
\code{pacfin_species_code = "SABL"}.}

\item{password}{Your password for the \code{datasourcename}.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}
}
\value{
RData frames is saved to the disk and the pulled data
are returned as a data frame.
}
\description{
Pull Catch Data from the Comprehensive Fish Ticket table
in the PacFIN database.
}
\author{
John R. Wallace, Kelli Faye Johnson
}
