% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullBDS.PacFIN.R
\name{PullBDS.PacFIN}
\alias{PullBDS.PacFIN}
\title{Pull biological data from the PacFIN database}
\usage{
PullBDS.PacFIN(
  pacfin_species_code,
  username = getUserName("PacFIN"),
  password = ask_password(),
  savedir = getwd(),
  verbose = TRUE
)
}
\arguments{
\item{pacfin_species_code}{A vector of strings specifying the PacFIN species
code(s) you are interested in. This has sometimes been referred to as
\code{"SPID"} in legacy sql scripts. An example for sablefish would be
\code{pacfin_species_code = "SABL"}. Lists of species codes in \href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/sp_tree.txt}{hierarchical order},
by
\href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/agency_spcodes.txt}{organization},
and \href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/agency_spcodes_spid.txt}{alphabetically organized}
can be found on the PacFIN website](https://pacfin.psmfc.org). Often you
will want to include nominal species categories. Where, nominal (i.e.,
\href{https://www.google.com/search?q=NOMINAL&rlz=1C1GCEJ_enUS866US866&sourceid=chrome&ie=UTF-8}{existing in name only})
means information for a given species that is \verb{derived' from non-species specific information, e.g., species complexes that are split out by species compositions like }nominal aurora rockfish' which would be \code{ARR1}. For some
functions, these nominal categories can automatically be added, see the
argument \code{addnominal}.}

\item{username}{The username you would like to use to access the database.
The default is to generate \code{username} from your computer name where
there are rules for each supported database. Sometimes this fails because
of legacy rules that are unknown to the development team. Please email the
maintainer of this package if you need more behind-the-scenes
functionality here.}

\item{password}{Your password for the database of interest. For some
functions, it may be okay to leave this argument missing because
you will be prompted for your password but not all. Please do not share
this password with anyone or push any code to a repository that has your
password saved.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors.
The printing doesn't affect any of the returned objects, instead,
printing makes it easier to see if the data has errors and what was
done to rectify them in the given function.
Sorry, but the default is to always print to the screen,
i.e., \code{verbose = TRUE}, so you do not miss out on anything.
This is standard practice for packages in the pfmc-assessments organization.}
}
\value{
An RData object is saved to the disk and the pulled data
are returned as a data frame.
The saved data can be read back in using \link{load}, but note that
the object is named \code{bds.pacfin} inside of the \code{.RData} file, and thus,
the object will retain this name within your work space.
In February of 2021, the returned object was renamed from \code{out}
to \code{bds.pacfin} in an attempt to be more clear.

Upon download, the data are changed from a long table to a wide table
using the combination of unique \code{FISH_ID} and \code{AGE_SEQUENCE_NUMBER}.
Users are warned if \code{verbose = TRUE} when there are non-unique \code{FISH_IDs} and these data
are automatically removed from the data frame.
This change from long to wide allows for rows equating to a single fish with columns
containing information about all measurements for that fish. Multiple age reads and
information about those reads such as age reader will be in the columns.
So, if your data had at most three duplicate reads,
then you will see numbers up to \code{3} pasted onto the end of relevant column names.
This is to help you match duplicate reads to other relevant information.
Not all double reads are currently available within PacFIN and
users should contact the ageing labs if they wish to inform ageing-error matrices.

\code{AGE_COUNT} is a somewhat cryptic column name and does not always make sense
when compared to \code{AGE_SEQUENCE_NUMBER}. It was determined that the former is
useful to identify how many potential agers were exposed to this fish.
For example, if \code{AGE_SEQUENCE_NUMBER} has a maximum value of three for a
given \code{FISH_ID}, then you can expect \code{AGE_COUNT} to be three for all three
rows in the PACFIN database for that fish. This is not always true though.
Sometimes, not all \code{AGE_SEQUENCE_NUMBER}s are present and they can skip
numbers for a given \code{FISH_ID}, and in this case, \code{AGE_COUNT} will be the
maximum \code{AGE_SEQUENCE_NUMBER} for a given \code{FISH_ID}.

\code{FINAL_FISH_AGE_IN_YEARS} is known as the best age for a given fish.
This will not always match an age reader or be a number determinable
from the individual age reads in \code{AGE_IN_YEARS}. Patrick explained to me
that when age reads do not agree, particularly for younger fish, then
the senior reader will work together with the junior reader to determine
an agreed-upon age. Other times, the senior reader's value will always
be used, or it could be that together they determine that they were both
wrong and a new age is proposed as the \verb{resolved age}. Nevertheless,
it can be quite messy and there is no way to predict the best age.
}
\description{
Pull biological data from the
Comprehensive Biological Data (comprehensive_bds_comm) table
in the PacFIN database (PACFIN_MARTS).
Data are in their raw form and should be cleaned either by hand or
using \link{cleanPacFIN}.
}
\examples{
\dontrun{
# Enter password in place of xxx below
mypw <- "xxx"
pd <- PullBDS.PacFIN(pacfin_species_code = "POP", password = mypw)
}

}
\seealso{
\itemize{
\item \link{cleanColumns} to change to legacy column names
\item \link{cleanPacFIN} to manipulate and subset the returned object
}
}
\author{
John R. Wallace, Kelli F. Johnson
}
