% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullCatch.PacFIN.R
\name{PullCatch.PacFIN}
\alias{PullCatch.PacFIN}
\title{Pull Catch Data from PacFIN Database}
\usage{
PullCatch.PacFIN(
  pacfin_species_code,
  username = getUserName("PacFIN"),
  password,
  savedir = getwd(),
  addnominal = TRUE
)
}
\arguments{
\item{pacfin_species_code}{A character string for the PacFIN
species code that you desire to pull data for. Known also as
\code{"SPID"} in the PacFIN database if using legacy sql scripts.
An example for sablefish would be
\code{pacfin_species_code = "SABL"}.}

\item{username}{The username you would like to use to access data from
\code{datasourcename}. The default is to generate \code{username}
from your computer name.}

\item{password}{Your password for the \code{datasourcename}.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}

\item{addnominal}{A logical, the default \code{TRUE} adds the nominal
PacFIN species code to the list of those searched for. \code{FALSE}
leaves \code{pacfin_species_code} as input by the user which will miss
catch for species such as dover sole that have more than one species
code within PacFIN, e.g., \code{pacfin_species_code = c("DOVR", "DVR1")}.
Users can also input a vector to pacfin_species_code if you want to
specify the species and the nominal species code without searching for
it. This is helpful for when you are getting data for two species and
you only want nominal catch for one or if you only wanted nominal catch.}
}
\value{
RData frames are saved to the disk and the pulled data
are returned as a data frame.
\itemize{
\item CompFT - pulled data
\item Research.Tribal.Catch - summary by year, state, and fleet;
fleets are as follows:
limited entry (LE), open access (OA), treaty indian (TI),
research (R), and unknown (XX)
\item Catch.INPFC - catch by INPFC area
\item Catch.PSMFC - catch by PSMFC area
}
}
\description{
Pull Catch Data from the Comprehensive Fish Ticket table
in the PacFIN database.
}
\author{
John R. Wallace, Kelli Faye Johnson
}
