% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullCatch.PacFIN.R
\name{PullCatch.PacFIN}
\alias{PullCatch.PacFIN}
\title{Pull catch data from PacFIN}
\usage{
PullCatch.PacFIN(
  pacfin_species_code,
  council_code = "P",
  username = getUserName("PacFIN"),
  password = ask_password(),
  savedir = getwd(),
  addnominal = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{pacfin_species_code}{A vector of strings specifying the PacFIN species
code(s) you are interested in. This has sometimes been referred to as
\code{"SPID"} in legacy sql scripts. An example for sablefish would be
\code{pacfin_species_code = "SABL"}. Lists of species codes in \href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/sp_tree.txt}{hierarchical order},
by
\href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/agency_spcodes.txt}{organization},
and \href{https://pacfin.psmfc.org/pacfin_pub/data_rpts_pub/code_lists/agency_spcodes_spid.txt}{alphabetically organized}
can be found on the \href{https://pacfin.psmfc.org}{PacFIN website}. Often you
will want to include nominal species categories. Where, nominal (i.e.,
\href{https://www.google.com/search?q=NOMINAL&rlz=1C1GCEJ_enUS866US866&sourceid=chrome&ie=UTF-8}{existing in name only})
means information for a given species that is "derived" from non-species
specific information, e.g., species complexes that are split out by species
compositions like "nominal aurora rockfish" which would be \code{ARR1}. For some
functions, these nominal categories can automatically be added, see the
argument \code{addnominal}.}

\item{council_code}{A vector of character strings specifying the council code
that you wish to retain data for. The default \code{"P"} means data will only be
returned for the Pacific Fisheries Management Council. Other accepted
values are \code{"N"} or \code{"*"}.}

\item{username}{Most often, this is a string containing your username for the
database of interest. You can use \code{\link[=getUserName]{getUserName()}} if you prefer to not
enter this argument and assume the default search and/or rules for finding
your username will work. This is the default behavior if you leave
\code{username} as a missing argument, i.e., \code{username <- getUserName(database = database)}. Sometimes this search will fail because of legacy rules, which
are unknown to the development team, that were used to create your
username. Please email the maintainer of this package if you need more
functionality here.}

\item{password}{Most often, this is a string containing your password for
the database of interest. You can use the function \code{\link[=ask_password]{ask_password()}} if you
would prefer to be prompted for your password. Please do not share this
password with anyone or push code to a repository that has your password
saved in it.}

\item{savedir}{A file path to the directory where the results will be saved.
The default is the current working directory.
The path can be relative or absolute.}

\item{addnominal}{A logical, where the default, \code{TRUE}, adds nominal
PacFIN species code to \code{pacfin_species_code}. \code{FALSE}
leaves \code{pacfin_species_code} as input by the user, which will miss
catch for species such as dover sole that have more than one species
code within PacFIN, e.g., \code{pacfin_species_code = c("DOVR", "DVR1")}.
Users can also input a vector to pacfin_species_code if you want to
specify the species and the nominal species code without searching for
it. This is helpful for when you are getting data for two species and
you only want nominal catch for one or if you only wanted nominal catch.
Nominal species code will be searched for using \code{\link[=PullNominal.PacFIN]{PullNominal.PacFIN()}} if
the input value for \code{addnominal} is \code{TRUE}.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
A \code{.RData} file is saved with the object inside the file stored as
\code{catch.pacfin}. This same data frame is also returned invisibly.
}
\description{
Pull catch data from the Comprehensive Fish Ticket table in PacFIN.
}
\details{
\subsection{Species with special considerations}{
\subsection{URCK}{

There is a rockfish (URCK) category that consists of unassigned rockfish
landings. The majority of the catch is prior to 2001. Currently, there is no
agreed upon methodology to parsed these landing out to specific rockfish
species. At present, landings in this category are not included in
species-specific rockfish catch pulls.
}

\subsection{POP}{

In PacFIN there are four species code that can be associated with Pacific
ocean perch. These are POP, POP1, POP2, UPOP. The POP1 is general shelf/slope
rockfish and not Pacific ocean perch specific landings. These records occur
only in Oregon. As of the 2017 assessment, these records should be removed
from the catch file.
}

}

\subsection{Searching for species}{

Values passed to \code{pacfin_species_code} are searched for in their exact form.
That is, there are no regular expression searches so mistakes such as
\code{" POP"} will not be found. In my experiences these mistakes in the species
codes are more common for PacFIN species codes that are three letters rather
than the standard four letters.
}
}
\examples{
\dontrun{
catch.pacfin <- PullCatch.PacFIN("PTRL")
# Check for confidentiality by year
# though you would also want to do this by your gear types
dplyr::group_by(catch.pacfin, LANDING_YEAR) \%>\%
  dplyr::summarize(count = dplyr::n_distinct(VESSEL_ID)) \%>\%
  dplyr::filter(count < 4)

# look for foreign landings
catch.pacfin <- PullCatch.PacFIN("PTRL", council_code = c("*", "N"))
# Counts of NROW() by area code(s)
dplyr::group_by(catch.pacfin, ORIG_PACFIN_CATCH_AREA_CODE) \%>\%
  dplyr::count()
}

}
\seealso{
\itemize{
\item \code{\link[=PullNominal.PacFIN]{PullNominal.PacFIN()}} determines the nominal species codes
\item \code{\link[=sql_catch]{sql_catch()}} writes the sql code to pull the data
}
}
\author{
Kelli F. Johnson
}
