% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullNominal.PacFIN.R
\name{PullNominal.PacFIN}
\alias{PullNominal.PacFIN}
\title{Pull the species information table and return nominal code}
\usage{
PullNominal.PacFIN(pacfin_species_code, username, password)
}
\arguments{
\item{pacfin_species_code}{A character string for the PacFIN
species code that you desire to pull data for. Known also as
\code{"SPID"} in the PacFIN database if using legacy sql scripts.
An example for sablefish would be
\code{pacfin_species_code = "SABL"}.}

\item{username}{The username you would like to use to access data from
\code{datasourcename}. The default is to generate \code{username}
from your computer name.}

\item{password}{Your password for the \code{datasourcename}.}
}
\value{
A vector of character values, where each value is a
PACFIN_SPECIES_CODE that pertains to the input \code{pacfin_species_code}.
\code{NA} is returned if no values are found.
}
\description{
Pull the species information table with PACFIN_SPECIES_CODE and
PACFIN_COMMON_NAME from PacFIN to determine the nominal code for
a given species code.
}
