% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stratify.R
\name{Stratify}
\alias{Stratify}
\title{Create a vector of stratifications}
\usage{
Stratify(inVector = NULL, splits = NULL, Nnames = NULL, numeric = F)
}
\arguments{
\item{inVector}{Input vector of values.  May be numeric or character.}

\item{splits}{List of values to use to stratify the input vector.  May be
character literals or numeric ranges.}

\item{Nnames}{Vector the length of \code{splits} naming the strata.}

\item{numeric}{A logical value, where if \code{TRUE} then
\code{findInterval} will be used and stratum 0 will be assigned to values
smaller than the first value in splits.}
}
\value{
Returns a vector the length of the input vector, with either numeric 
levels indicating the stratification of the input, or with Nnames assigning
levels of the input vector.
}
\description{
\code{Stratify} takes an input vector and list of values used to designate
the values returned in the strats vector.
}
\details{
\subsection{Workflow}{
\code{Stratify} is intended to add levels for stratification before
running \code{\link{getComps}}.
}


Any values not given in \code{splits} will be assigned to stratum 0.
This function can be used to designate fleet or for stratification
based on depth or INPFC area.
}
\examples{
Pdata$use_depth = Stratify( Pdata$DEPTH_AVG, splits=list(0, 50, 100, 250, 500), numeric=T)

Pdata$my_fleets = Stratify( Pdata$GEAR,
           splits=list("GFS", c("MDT","TB"), "TR", c("LGL","FTS")),
           Nnames=list("G1","G2","G3","G4") )

}
\author{
Andi Stephens
}

