% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAge.R
\name{summaryAgeMethod}
\alias{summaryAgeMethod}
\title{Summary of available age methods}
\usage{
summaryAgeMethod(Pdata, verbose = TRUE)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{verbose}{A logical specifying if output should be written to the
screen or not. Good for testing and exploring your data but can be turned
off when output indicates information that you already know. The printing
of output to the screen does not affect any of the returned objects. The
default is to always print to the screen, i.e., \code{verbose = TRUE}.}
}
\value{
A data frame of counts for each age method.
If \code{verbose}, then text explaining the summary is also printed to the
screen prior to returning the table.
}
\description{
Use \code{\link[dplyr:count]{dplyr::count()}} to summarize the available age types.
}
