% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAge.R
\name{summaryAgeMethod}
\alias{summaryAgeMethod}
\title{Summary of available age methods}
\usage{
summaryAgeMethod(Pdata, verbose = FALSE)
}
\arguments{
\item{Pdata}{A data frame of biological samples
originating from the
Pacific Fishieries Information Network (PacFIN) data warehouse,
which originated in 2014. Data are pulled using sql calls, see
\code{\link{PullBDS.PacFIN}()}.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors.
The printing doesn't affect any of the returned objects, instead,
printing makes it easier to see if the data has errors and what was
done to rectify them in the given function.
Sorry, but the default is to always print to the screen,
i.e., \code{verbose = TRUE}, so you do not miss out on anything.
This is standard practice for packages in the nwfsc-assess organization.}
}
\value{
A data frame of counts for each age method.
If \code{verbose}, then text explaining the summary is also printed to the
screen prior to returning the table.
}
\description{
Use \link[dplyr:count]{dplyr::count} to summarize the available age types.
}
