% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeComps.R
\name{writeComps}
\alias{writeComps}
\title{Write out composition data formatted for Stock Synthesis.}
\usage{
writeComps(inComps, fname = "out.csv", abins = NULL, lbins = NULL,
  maxAge = Inf, partition = 0, ageErr = 0, out = "FthenM",
  dummybins = FALSE, verbose = TRUE)
}
\arguments{
\item{inComps}{A dataframe generated as described in Workflow, above.}

\item{fname}{A filename, used to save the output as a \code{.csv} file.
  Users can specify a full path if they do not want the file written in the
  current working directory.}

\item{abins}{Bins to use for ages.  Default is the data bins.}

\item{lbins}{Bins to use for lengths.  Default is the data bins.}

\item{maxAge}{A numeric value specifying the maximum age of fish that
  should be included in the composition data, unless \code{dummybins = TRUE},
  then those fish will be included in a plus group which you can investigate
  later. Note that \code{maxAge} is only used if \code{abins = NULL}, otherwise
  fish are binned according to user specified bins irregardless of \code{maxAge}.}

\item{partition}{Defaults to 1.}

\item{ageErr}{Defaults to 1.}

\item{out}{One of ("FthenM", "Fout", "Mout", or "Uout").}

\item{dummybins}{A logcial value specifying whether data outside of the
lower and upper \code{abins} or \code{lbins} should be added to dummy bins,
or be placed in the specified bins. Default is \code{TRUE}.  Dummy
bins are useful for determining whether the current bin structure
properly captures the data.}

\item{verbose}{A logical value specifying if output should be written
to the screen or not. Good for testing and exploring your data but
can be turned off when output indicates errors. The default value
is \code{TRUE}.}
}
\value{
Appends data to the file given in \code{fname}.

  Conditionally returns a dataframe if the \code{out} argument
  specifies the type of composition to return.
}
\description{
Write out composition data to a file, binning the data as specified.

\subsection{Workflow}{
PacFIN data should first be stratified using \code{\link{getComps}},
then sexes should be assigned as males and females by \code{\link{doSexRatio}}
before using \code{writeComps}.

Failure to use \code{doSexRatio} will result in all unsexed fish being discarded.
}
}
\details{
The structure of the input dataframe determines whether
  \code{writeComps} produces age-, length-, or conditional-age-at-length-
  composition data.

  Four sets of composition data are written to a single file specified
  by \code{fname}.  These are:
  \itemize{
  \item{females followed by males}
  \item{male data only}
  \item{female data only}
  \item{males and females recombined as unsexed fish}
  }

  The output file is appended to, rather than overwritten, so you may
  want to specify a new filename each time you generate a different
  stratification or bin structure, e.g., "out.1.csv", "out.2.csv".

  Composition data are raw weights rather than proportions. Stock Synthesis
  internally converts these to proportions.  The raw weights should be
  examined for anomalies.

  To create proportions use \code{prop.table} on the columns
  containing composition data.

  The columns in the output preceeding 'lengthcm' or 'age' are those that
  were used in stratifying the data.

\subsection{Reality Checks}{
\itemize{
\item{Set \code{verbose = TRUE} to follow progress}
\item{Use \code{dummybins} to evaluate bin structure}
\item{Examine the raw-weight output for anomalies}
}}
}
\author{
Andi Stephens
}
\seealso{
\code{\link{getComps}}, \code{\link{doSexRatio}}
}

